% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_cellType_gene_max.R
\name{a_cellType_gene_max}
\alias{a_cellType_gene_max}
\title{a_cellType_gene_max}
\usage{
a_cellType_gene_max(anno_interactions_allSamples, verbose = FALSE)
}
\arguments{
\item{anno_interactions_allSamples}{a list of data frames containing annotation information for interactions in all samples.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
a list of data frames (numeric \link{0,1}) with an additional column for the maximum active fraction for ligand in sending cell type and another for the maximum active fraction for receptor in receiving cell type.
}
\description{
calculates the maximum active fraction for ligand in sending cell type and receptor in receiving cell type for each gene over all samples.
}
\examples{
# load example data
data(example_anno_interactions_allSamples)

}
