% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_ct_g_allSamples.R
\name{a_ct_g_allSamples}
\alias{a_ct_g_allSamples}
\title{a_ct_g_allSamples}
\usage{
a_ct_g_allSamples(
  counts,
  anno_cells,
  anno_interactions_allSamples,
  threshold_expr = 0.05,
  threshold_nr_active_cells = 0,
  verbose = FALSE
)
}
\arguments{
\item{counts}{numeric dataframe: normalized expression data frame containing ligands and receptors in the rows and cells in the columns. Row names and column names should be defined.}

\item{anno_cells}{dataframe: concatenated data frame of the cell annotation from all samples (rows are cell IDs, columns annotation columns: must contain 'cell_ID', 'cell_type' and 'sample_ID' column).}

\item{anno_interactions_allSamples}{list: list containing the interactions for each sample.}

\item{threshold_expr}{numeric: expression threshold below which the expression of the ligand (receptor) is not considered, default is 0.05.}

\item{threshold_nr_active_cells}{numeric: threshold for minimal number of active cells, default is 0.}

\item{verbose}{logical: whether to print progress messages or not, default is FALSE.}
}
\value{
list:
\itemize{
\item anno_interactions_allSamples: list containing the interactions for each sample, with additional columns 'a_s_l', 'nr_s_l_active', 'a_r_r' and 'nr_r_r_active' indicating the active fraction and active number of cells for the ligand and receptor in the sending and receiving cell types respectively.
}
}
\description{
calculates the relative active fraction for each gene in each cell type for all samples.
}
\examples{
# load data
data("counts")
data("anno_cells")
data("anno_interactions_allSamples")


}
