% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{Document a dataset}
\usage{
add_data(path, package_path = "datapackage.json", write = TRUE)
}
\arguments{
\item{path}{A character vector of paths to tabular data files.}

\item{package_path}{Package to add the metadata to; path to the .json file, or a list with the read-in version.}

\item{write}{Logical; if \code{FALSE}, returns the \code{paths} metadata without reading or rewriting \code{package}.}
}
\value{
A version of the metadata, which is also added to \code{datapackage.json} if \code{write = TRUE}.
}
\description{
Add information about variables in a dataset to a \code{datapackage.json} metadata file.
}
\examples{
\dontrun{
# write example data
write.csv(mtcars, "mtcars.csv")

# add it to an existing datapackage.json file in the current working directory
add_data("mtcars.csv")
}
}
\seealso{
Initialize the datapackage.json file with \code{\link{init_data}}.
}
