% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDB.R
\name{annotate_components}
\alias{annotate_components}
\title{Annotate Components}
\usage{
annotate_components(complete_data)
}
\arguments{
\item{complete_data}{A data frame with columns 'Ligand' and 'Receptor'.}
}
\value{
A data frame with columns 'genesymbol', 'score', and 'parent', where 'parent' indicates the parent category of the gene
}
\description{
This function annotates components (genes) from the PPI network with their parent category. It uses data from OmniPath to determine if the components are categorized as ligands or receptors, and handles cases where the components fall into other categories.
}
\examples{
mock <- data.frame(Ligand = c("IL10","IGF1"), Receptor = c("JAK2","ITGA8"))
annotated_data <- annotate_components(mock)

}
