% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDB.R
\name{auto_update_db}
\alias{auto_update_db}
\title{Auto Update Database}
\usage{
auto_update_db(db_type)
}
\arguments{
\item{noncurated}{Logical, if TRUE processes the non-curated database.}

\item{curated}{Logical, if TRUE processes the curated database.}

\item{both}{Logical, if TRUE processes both the non-curated and curated databases.}
}
\value{
Returns the updasted database.
}
\description{
This function automatically updates the database by processing
non-curated, curated, or both datasets.
}
\examples{
updated_db <- auto_update_db("noncurated")
}
