% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_access.R
\name{calculate_access}
\alias{calculate_access}
\title{Calculate spacial access scores.}
\usage{
calculate_access(consumers, providers, cost = NULL, buffers = NULL,
  selection_weight = FALSE, consumers_id = "geoid",
  consumers_value = "count", providers_id = "geoid",
  providers_value = "count")
}
\arguments{
\item{consumers}{Number of consumers (demand); either a vector with consumer amounts (such as population counts),
or a matrix-like object with a column of IDs (located by \code{consumers_id}) and a column of amounts (located by
\code{consumers_value}).}

\item{providers}{Number of providers (supply); either a vector with provider amounts (such as number of doctors),
or a matrix-like object with a column of IDs (located by \code{providers_id}) and a column of amounts (located by
\code{providers_value}).}

\item{cost}{Cost associated with each pair of \code{consumers} and \code{providers}, such as distance or travel
time.}

\item{buffers}{Means of associating \code{consumers} to \code{providers}. The simplest is a single number
representing a maximum distance from each provider, effectively selecting only consumers within
a catchment area (2-step floating catchment area). An enhancement of this is a list of vectors with two values
each: the first is a distance, and the second a weight to associate with that distance (e.g.,
\code{list(c(10, 1), c(20, .5))}, which will give consumers within a \code{cost} of 10 full weight, and those within
a \code{cost} of 20 half weight; enhanced 2-step floating catchment area).}

\item{selection_weight}{Logical; if \code{TRUE}, will adjust the weight of \code{buffers} by a value associated
with each consumer (3-step floating catchment area).}

\item{consumers_id, consumers_value, providers_id, providers_value}{Column names in \code{consumers} and/or
\code{providers} to extract IDs and values.}
}
\value{
A vector with an access score for each \code{consumer}.
}
\description{
Calculate variants of floating catchment areas -- measures of access as potentially weighted consumer to provider
ratios.
}
\examples{
pop <- c(5, 10, 50)
doc <- c(50, 100)
travel_time <- matrix(c(5, 50, 25, 70, 40, 30), ncol = 2)

# 2-step floating catchment area
calculate_access(pop, doc, travel_time, 30)

# enhanced 2-step floating catchment area
calculate_access(pop, doc, travel_time, list(c(60, .22), c(40, .68), c(20, 1)))

# 3-step floating catchment area
calculate_access(pop, doc, travel_time, list(c(60, .22), c(40, .68), c(20, 1)), TRUE)
}
\references{
Lou, W. & Qi, Y. (2009). An enhanced two-step floating catchment area (E2SFCA) method for measuring spatial
accessibility to primary care physicians. \emph{Health & Place, 15}, 1100-1107.
doi: \href{https://doi.org/10.1016/j.healthplace.2009.06.002}{10.1016/j.healthplace.2009.06.002}

Radke, J. & Mu, L. (2000). Spatial decompositions, modeling and mapping service regions to predict access
to social programs. \emph{Geographic Information Sciences, 6}, 105-112.
doi: \href{https://doi.org/10.1080/10824000009480538}{10.1080/10824000009480538}

Wan, N., Zou, B., & Sternberg, T. (2012). A three-step floating catchment area method for analyzing spatial
access to health services. \emph{International Journal of Geographical Information Science, 26}, 1073-1089.
doi: \href{https://doi.org/10.1080/13658816.2011.624987}{10.1080/13658816.2011.624987}
}
