% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_communication.R
\name{calculate_communication}
\alias{calculate_communication}
\title{calculate_communication}
\usage{
calculate_communication(
  counts,
  anno_samples,
  anno_cells,
  threshold_expr = 0.05,
  threshold_nr_active_cells = 0,
  threshold_celltype_size = 4,
  lrp_database,
  verbose = FALSE
)
}
\arguments{
\item{counts}{numeric dataframe: normalized expression data frame containing ligands and receptors in the rows and cells in the columns. Row names and column names should be defined.}

\item{anno_samples}{dataframe: concatenated data frame of the sample annotation from all samples (rows are sample IDs, columns annotation columns: must contain "sample_ID" and "case_or_control" columns).}

\item{anno_cells}{dataframe: concatenated data frame of the cell annotation from all samples (rows are cell IDs, columns annotation columns: must contain "cell_ID", "cell_type" and "sample_ID" columns).}

\item{threshold_expr}{numeric: expression threshold below which the expression of the ligand (receptor) is not considered. The default value is 0.05.}

\item{threshold_nr_active_cells}{numeric: threshold for minimal number of active cells.The default value is 0.}

\item{threshold_celltype_size}{numeric: threshold of minimum number of cells in a cell type in a sample. The default value is 4.}

\item{lrp_database}{dataframe: contains L and R Pairs, should contain solumns Ligand.ApprovedSymbol, Receptor.ApprovedSymbol and Pair.Name.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
list of the following objects:
\itemize{
\item per_sample_anno_interactions a list of dataframes (one dataframe per sample). The names of the dataframes are equal to sample_ID. Each dataframe contains the interaction IDs as row names and the following columns:
-- interaction_ID: an interaction ID is constructed in the following form: "sending cell type : ligand _ receiving cell type : receptor". E.g. in an interaction "T:TGFB1_B:CXCR4", the T cells are the sending cell type expressing the TGFB1 ligand and the B cells are the receiving cell type expressing the CXCR4 receptor.
-- ligand_gene_name: the gene symbol of the ligand, e.g. "TGFB1"
-- receptor_gene_name: the gene symbol of the receptor, e.g. "CXCR4"
-- sending_cell_type: the sending cell type, e.g. "T"
-- receiving_cell_type: the receiving cell type, e.g. "B"
-- f_s: numeric \link{0,1} the abundance (as fraction) of the sending cell type within the sample of interest
-- f_r: numeric \link{0,1} the abundance (as fraction) of the receiving cell type within the sample of interest
-- f_s_max: numeric \link{0,1} the maximum abundance (as fraction) value for the sending cell type over all samples in the dataset
-- f_r_max: numeric \link{0,1} the maximum abundance (as fraction) value for the receiving cell type over all samples in the dataset
-- a_s_l: numeric \link{0,1} the fraction of the sending cells expressing the ligand above the threshold in the sample of interest
-- nr_s_l_active: numeric \link{0,Inf} the number of the sending cells expressing the ligand above the threshold in the sample of interest
-- a_r_r: numeric \link{0,1} the fraction of the receiving cells expressing the receptor above the threshold in the sample of interest
-- nr_r_r_active: numeric \link{0,Inf} the number of the receiving cells expressing the receptor above the threshold in the sample of interest
-- a_s_l_max: numeric \link{0,1} the maximum fraction of the sending cells expressing the ligand above the threshold over all samples
-- a_r_r_max: numeric \link{0,1} the maximum fraction of the receiving cells expressing the receptor above the threshold over all samples
-- e_s_l: numeric \link{0,Inf} the mean expression of the ligand in the active fraction of the sending cells in the sample of interest
-- e_r_r: numeric \link{0,Inf} the mean expression of the receptor in the active fraction of the receiving cells in the sample of interest
-- e_s_l_max: numeric \link{0,Inf} the maximum mean expression of the ligand in the active fraction of the sending cells over all samples
-- e_r_r_max: numeric \link{0,Inf} the maximum mean expression of the receptor in the active fraction of the receiving cells over all samples
-- rho_s: numeric \link{0,1} the normalized abundance of the sending cell type within the sample of interest
-- rho_r: numeric \link{0,1} the normalized abundance of the receiving cell type within the sample of interest
-- rho: numeric \link{0,1} the normalized abundance component of the weight formula
-- phi_s_l: numeric \link{0,1} the normalized fraction of the sending cells expressing the ligand above the threshold in the sample of interest
-- phi_r_r: numeric \link{0,1} the normalized fraction of the receiving cells expressing the receptor above the threshold in the sample of interest
-- phi: numeric \link{0,1} the normalized active fraction component of the weight formula
-- p_s_l: numeric \link{0,1} the normalized mean expression of the ligand in the active fraction of the sending cells in the sample of interest
-- p_r_r: numeric \link{0,1} the normalized mean expression of the receptor in the active fraction of the receiving cells in the sample of interest
-- p: numeric \link{0,1} the normalized mean expression in the active fraction component of the weight formula
-- w: numeric \link{0,1} the interaction weight
\item f_s numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being abundance (as fraction) of the sending cell type
\item f_r numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being abundance (as fraction) of the receiving cell type
\item rho_s numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being normalized abundance of the sending cell type
\item rho_r numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being normalized abundance of the receiving cell type
\item rho numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being the normalized abundance component of the weight formula
\item a_s_l numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being the fraction of the sending cells expressing the ligand above the threshold
\item a_r_r numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being the fraction of the receiving cells expressing the receptor above the threshold
\item nr_s_l_active numeric dataframe \link{0,Inf} with row names being interaction IDs, column names being sample IDs and values being the number of the sending cells expressing the ligand above the threshold
\item nr_r_r_active numeric dataframe \link{0,Inf} with row names being interaction IDs, column names being sample IDs and values being the number of the receiving cells expressing the receptor above the threshold
\item phi_s_l numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being the normalized fraction of the sending cells expressing the ligand above the threshold
\item phi_r_r numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being the normalized fraction of the receiving cells expressing the receptor above the threshold
\item phi numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being the normalized fraction of active cells component of the weight formula
\item e_s_l numeric dataframe \link{0,Inf} with row names being interaction IDs, column names being sample IDs and values being the mean expression of the ligand in the active fraction of the sending cells
\item e_r_r numeric dataframe \link{0,Inf} with row names being interaction IDs, column names being sample IDs and values being the mean expression of the receptor in the active fraction of the receiving cells
\item p_s_l numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being the normalized mean expression of the ligand in the active fraction of the sending cells
\item p_r_r numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being the normalized mean expression of the receptor in the active fraction of the receiving cells
\item p numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being the normalized mean expression in the active fraction component of the weight formula
\item weights numeric dataframe \link{0,1} with row names being interaction IDs, column names being sample IDs and values being the interaction weight
\item thresholds: a list of:
-- threshold_expr numeric: expression threshold below which the expression of the ligand (receptor) is not considered
-- threshold_nr_active_cells: numeric: threshold for minimal number of active cells
-- threshold_celltype_size: numeric: threshold of minimum number of cells in a cell type in a sample
\item anno_interactions: a list of:
-- interaction_ID: an interaction ID is constructed in the following form: "sending cell type : ligand _ receiving cell type : receptor". E.g. in an interaction "T:TGFB1_B:CXCR4", the T cells are the sending cell type expressing the TGFB1 ligand and the B cells are the receiving cell type expressing the CXCR4 receptor.
-- ligand_gene_name: the gene symbol of the ligand, e.g. "TGFB1"
-- receptor_gene_name: the gene symbol of the receptor, e.g. "CXCR4"
-- sending_cell_type: the sending cell type, e.g. "T"
-- receiving_cell_type: the receiving cell type, e.g. "B"
\item anno_samples dataframe: concatenated data frame of the sample annotation from all samples (rows are sample IDs, columns annotation columns: must contain "sample_ID" and "case_or_control" columns).
\item anno_cells dataframe: concatenated data frame of the cell annotation from all samples (rows are cell IDs, columns annotation columns: must contain "cell_ID", "cell_type" and "sample_ID" columns).
}
}
\description{
calculates communication between cell types based on the provided count data and annotation information.
}
\examples{
counts <- data.frame(
      gene = c("gene1","gene2","gene3")
   ,cell1 = c(1,2,3)
 ,cell2 = c(4,5,6)
)
rownames(counts) <- counts$gene
counts$gene <- NULL
anno_cells <- data.frame(
     cell_ID = c("cell1","cell2")
  ,cell_type = c("cell_type1","cell_type2")
)
anno_samples <- data.frame(
     sample_ID = c("sample1")
 ,health_status = c("healthy")
,case_or_control = c("case")
)
lrp_database <- read.csv("LRP_database.csv")
calculate_communication(counts = counts
                      ,anno_samples = anno_samples
                     ,anno_cells = anno_cells
                    ,lrp_database = lrp_database
                  ,verbose = TRUE
)

}
