% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_weight.R
\name{calculate_weight}
\alias{calculate_weight}
\title{calculate_weight}
\usage{
calculate_weight(anno_interactions, verbose = FALSE)
}
\arguments{
\item{anno_interactions}{dataframe: contains annotations to the interactions in the correspoding sample. The dataframe contains interactions IDs as row names, as well as the following columns: interaction_ID, ligand_gene_name, receptor_gene_name, sending_cell_type, receiving_cell_type, f_s, f_s_max, f_r, f_r_max, a_s_l, a_s_l_max, a_r_r, a_r_r_max, e_s_l, e_s_l_max, e_r_r, e_r_r_max.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
anno_interactions dataframe with the following columns added: rho_s, rho_r, rho, phi_s_l, phi_r_r, phi, p_s_l, p_r_r, p, w.
}
\description{
calculates individual components and the weight per sample.
}
\examples{
anno_interactions <- data.frame(
       interaction_ID = c('interaction1','interaction2','interaction3')
     ,ligand_gene_name = c('gene1','gene2','gene3')
    ,receptor_gene_name = c('gene1','gene2','gene3')
  ,sending_cell_type = c('cellType1','cellType2','cellType3')
,receiving_cell_type = c('cellType1','cellType2','cellType3')
,mean_expr_l_s = c(0.1,0.2,0.3)
,mean_expr_r_r = c(0.1,0.2,0.3)
,max_mean_expr_all_samples = c(0.1,0.2,0.3)
,phi_l_s = c(0.1,0.2,0.3)
,phi_r_r = c(0.1,0.2,0.3)
)
anno_interactions <- calculate_weight(anno_interactions = anno_interactions
                                      ,verbose = TRUE
)

}
