% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_repository.R
\name{check_repository}
\alias{check_repository}
\title{Check Data Repositories}
\usage{
check_repository(dir = ".", search_pattern = "\\\\.csv(?:\\\\.[gbx]z2?)?$",
  value = "value", value_name = "measure", id = "geoid", time = "year",
  dataset = "region_type", entity_info = c("region_type", "region_name"),
  attempt_repair = FALSE)
}
\arguments{
\item{dir}{Root directory of the data repository.}

\item{search_pattern}{Regular expression used to search for data files.}

\item{value}{Name of the column containing variable values.}

\item{value_name}{Name of the column containing variable names.}

\item{id}{Column name of IDs which uniquely identify entities.}

\item{time}{Column name of the variable representing time.}

\item{dataset}{Column name used to separate data info sets (such as by region).}

\item{entity_info}{A vector containing variable names to go into making \code{entity_info.json}.}

\item{attempt_repair}{Logical; if \code{TRUE}, will attempt to fix some warnings in data files.}
}
\value{
An invisible list of check results.
}
\description{
Performs a series of checks to see if data in a given repository can be ingested by a datacommons project.
}
\examples{
\dontrun{
# from a data repository
check_repository()
}
}
