% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_template.R
\name{check_template}
\alias{check_template}
\title{Check the structure of a templated project}
\usage{
check_template(template, name = "", dir = ".", spec = NULL)
}
\arguments{
\item{template}{Name of the template (e.g., \code{"function"}).}

\item{name}{Name of the template instance (e.g., \code{"test_function"}).}

\item{dir}{Path to the base directory of the template.}

\item{spec}{The template's spec; specify if template spec files do not exist.}
}
\value{
A list with results of the check:
\tabular{rl}{
\strong{dir} \tab Path of the checked directory. \cr
\strong{files} \tab Path(s) of the checked files. \cr
\strong{spec} \tab A list with the template's spec. \cr
\strong{status} \tab A named logical vector indicating whether each components exist. \cr
\strong{incomplete} \tab A character vector with any existing files that still have template text. \cr
\strong{exists} \tab \code{all(status)} \cr
\strong{message} \tab A character vector including messages associated with failures. \cr
}
}
\description{
Check if the parts of a template are in place, based on the template's spec.
}
\examples{
\dontrun{
# precheck before a new template is made
check_template("template_new", list(name = "new", context = "new", files = list("new.R")))

# check the new template after creation
check_template("new")
}
}
