% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_lrp_database.R
\name{clean_lrp_database}
\alias{clean_lrp_database}
\title{clean_lrp_database}
\usage{
clean_lrp_database(counts, lrp_database, verbose = FALSE)
}
\arguments{
\item{counts}{dataframe: expression data frame containing ligands and receptors in the rows and cells in the columns. Rownames and column names should be defined.}

\item{lrp_database}{dataframe: contains L and R Pairs, should contain solumns Ligand.ApprovedSymbol, Receptor.ApprovedSymbol and Pair.Name.}

\item{verbose}{logical: whether to print the progress of the function or not.}
}
\value{
Dataframe: Extract L and R matched in counts for the analysis without duplicates and NA
}
\description{
clean_lrp_database
}
\examples{
counts <- data.frame(
      gene = c('gene1','gene2','gene3')
   ,cell1 = c(1,2,3)
 ,cell2 = c(4,5,6)
)
rownames(counts) <- counts$gene
counts$gene <- NULL
lrp_database <- read.csv('LRP_database.csv')
clean_lrp_database(counts
                 ,lrp_database
               ,verbose = TRUE
)

}
