% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\docType{package}
\name{community-package}
\alias{community}
\alias{community-package}
\title{community: Infer interactions between cell types}
\description{
Community is an R package for analyzing single-cell RNA sequencing data to infer interactions between cell types. The package includes functions for preprocessing and quality control, as well as functions for inferring interactions and for analyzing the inferred interactions.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/SoloveyMaria/community/issues}
}

}
\author{
\strong{Maintainer}: Maria Solovey \email{maria.solovey@bmc.med.lmu.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
