% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_initial_anno_interactions.R
\name{create_initial_anno_interactions}
\alias{create_initial_anno_interactions}
\title{create_initial_anno_interactions}
\usage{
create_initial_anno_interactions(
  counts,
  anno_samples,
  anno_cells,
  lrp_database,
  verbose = FALSE
)
}
\arguments{
\item{counts}{numeric \link{0,Inf}: normalized expression data frame contatining ligands and receptors in the rows and cells in the columns. Rownames and column names should be defined.}

\item{anno_samples}{dataframe: concatenated data frame of the sample annotation from all samples (rows are sample IDs, columns annotation columns: must contains "sample_ID" and case_or_control columns).}

\item{anno_cells}{dataframe: concatenated data frame of the cell annotation from all samples (rows are cell IDs, columns annotation columns: must contains "cell_ID", "cell_type" and "sample_ID" column).}

\item{lrp_database}{dataframe: contains L and R Pairs, should contain solumns Ligand.ApprovedSymbol, Receptor.ApprovedSymbol and Pair.Name.}
}
\value{
anno_interactions_allSamples: list of dataframes. The list has a length equal to the number of samples in the dataset, and has names equal to the sample IDs:
\itemize{
\item Each dataframe is annotation to the interactions in the correspoding sample.
\item Each dataframe contains interactions IDs in the rows
\item Each dataframe comtains following columns:
interaction_ID: A unique identifier for each interaction created by concatenating sending cell type, ligand gene name, receiving cell type, and receptor gene name.
ligand_gene_name: The gene name of the ligand in the interaction.
receptor_gene_name: The gene name of the receptor in the interaction.
sending_cell_type: The cell type that expresses the ligand.
receiving_cell_type: The cell type that expresses the receptor.
}
}
\description{
creates initial anno_interactions for all samples.
}
\examples{
     # load data
    data("counts")
   data("anno_samples")
  data("anno_cells")
data("lrp_database")
     # run function
   anno_interactions_allSamples <- create_initial_anno_interactions(counts
                                                   ,anno_samples
                                                  ,anno_cells
                                                ,lrp_database
  )
    # check output
  print("str(anno_interactions_allSamples)")
}
