% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_add.R
\name{data_add}
\alias{data_add}
\title{Adds documentation of a dataset to a datapackage}
\usage{
data_add(path, meta = list(), package_path = "datapackage.json",
  dir = ".", write = TRUE, refresh = FALSE, sha = "512")
}
\arguments{
\item{path}{A character vector of paths to plain-text tabular data files.}

\item{meta}{Information about each data file. A list with a list entry for each entry in \code{path}; see details.}

\item{package_path}{Package to add the metadata to; path to the .json file, or a list with the read-in version.}

\item{dir}{Directory in which to look for \code{path}, and write \code{package_path}.}

\item{write}{Logical; if \code{FALSE}, returns the \code{paths} metadata without reading or rewriting
\code{package_path}.}

\item{refresh}{Logical; if \code{TRUE}, will remove any existing dataset information.}

\item{sha}{A number specifying the Secure Hash Algorithm function,
if \code{openssl} is available (checked with \code{Sys.which('openssl')}).}
}
\value{
An invisible version of the metadata, which is also added to \code{datapackage.json} if \code{write = TRUE}.
}
\description{
Add information about variables in a dataset to a \code{datapackage.json} metadata file.
}
\details{
\code{meta} should be a list with unnamed entries for entry in \code{path}, and each entry can include a named entry
for any of these:
\describe{
\item{source}{
A list or list of lists with entries for at least \code{name}, and ideally for \code{url}.
}
\item{ids}{
A list or list of lists with entries for at least \code{variable} (the name of a variable in the dataset).
Might also include \code{name} (for a display name), and \code{value} with a parsing scheme. A parsing
scheme is either a regular expression (e.g., \code{'^.*\\\\.'} for the first part of an id, before a period)
or a string range (e.g., \code{c(0, 6)} for the first 6 characters of an id).
}
\item{time}{
A string giving the name of a variable in the dataset representing a repeated observation of the same entity.
}
\item{variables}{
A list with named entries providing more information about the variables in the dataset.
}
}
}
\examples{
\dontrun{
# write example data
write.csv(mtcars, "mtcars.csv")

# add it to an existing datapackage.json file in the current working directory
data_add("mtcars.csv")
}
}
\seealso{
Initialize the \code{datapackage.json} file with \code{\link{init_data}}.
}
