% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reformat_sdad.R
\name{data_reformat_sdad}
\alias{data_reformat_sdad}
\title{Reformat an SDAD-formatted dataset}
\usage{
data_reformat_sdad(files, value = "value", value_name = "measure",
  id = "geoid", time = "year", dataset = "region_type",
  value_info = "measure_type", entity_info = c(type = "region_type", name =
  "region_name"), formatters = NULL, out = NULL, compression = "xz",
  read_existing = FALSE, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{files}{A character vector of file paths, or the path to a directory containing data files.}

\item{value}{Name of the column containing variable values.}

\item{value_name}{Name of the column containing variable names; assumed to be a single variable per file if
not present.}

\item{id}{Column name IDs which uniquely identify entities.}

\item{time}{Column name of the variable representing time.}

\item{dataset}{Column name used to separate entity scales.}

\item{value_info}{A vector of column names which provide additional information about values
(such as their name and type).}

\item{entity_info}{A list containing variable names to extract and create an ids map from (
\code{entity_info.json}, created in the output directory). Entries can be named to rename the
variables they refer to in entity features.}

\item{formatters}{A list of functions to pass columns through, with names identifying those columns
(e.g., \code{list(region_name = function(x) sub(",.*$", "", x))} to strip text after a comma in the
"region_name" column).}

\item{out}{Path to a directory to write files to; if not specified, files will not be written.}

\item{compression}{A character specifying the type of compression to use on the created files,
between \code{"gzip"}, \code{"bzip2"}, and \code{"xz"}. Set to \code{FALSE} to disable compression.}

\item{read_existing}{Logical; if \code{TRUE}, will read in existing sets that are not to be overwritten.
Useful if you use the returned sets.}

\item{overwrite}{Logical; if \code{TRUE}, will overwrite existing reformatted files, even if
the source files are older than it.}

\item{verbose}{Logical; if \code{FALSE}, will not print status messages.}
}
\value{
An invisible list of the unified variable datasets, split into datasets.
}
\description{
Unify multiple files, which each contain a tall set of variables associated with regions.
}
\details{
The basic assumption is that there are (a) entities which (b) exist in a hierarchy, and
(c1) have a static set of features and (c2) a set of variable features which
(d) are assessed at multiple time points.

For example (and generally), entities are (a) regions, with (b) smaller regions making up larger regions,
and which (c1) have names, and (c2) population and demographic counts (d) between 2009 and 2019.
}
\examples{
\dontrun{
data_reformat_sdad("data/directory")
}
}
