% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacommons_map_variables.R
\name{datacommons_map_variables}
\alias{datacommons_map_variables}
\title{Map Data Commons Variables}
\usage{
datacommons_map_variables(dir,
  search_pattern = "(?:\\\\d{4}).+\\\\.csv\\\\.[gbx]z2?$",
  variable_location = "measure", sources = c("repos", "cache"),
  reader = read.csv, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{dir}{Directory of the data commons projects.}

\item{search_pattern}{A regular expression string used be passed to \code{\link{list.files}}.}

\item{variable_location}{The name of a column contain variable names in each dataset, or a function to retrieve
variable names (e.g., \code{colnames}).}

\item{sources}{Sub-directories of the data commons project to search for files in. If multiple are included,
prefers files from those in the earlier locations in the case of duplicates. Default is to prefer files
in the primary repository.}

\item{reader}{A function capable of handling a connection in its first argument, which returns a matrix-like object.}

\item{overwrite}{Logical; if \code{TRUE}, creates a new map even if one exists.}

\item{verbose}{Logical; if \code{FALSE}, does not print status messages.}
}
\value{
An invisible \code{data.frame} of the mapped variables, with \code{variable} (variable name),
\code{repo} (the repository containing the file), \code{full_name} (variable name with a prefix from the file's name),
and \code{file} (path to the file) columns.
}
\description{
Extract variables from files in datacommons repositories
}
\examples{
\dontrun{
# from a data commons project directory
map <- datacommons_map_variables(".")
}
}
