% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacommons_refresh.R
\name{datacommons_refresh}
\alias{datacommons_refresh}
\title{Refresh Data Commons Repositories}
\usage{
datacommons_refresh(dir, clone_method = "http",
  include_distributions = TRUE, only_new = FALSE, quiet = FALSE)
}
\arguments{
\item{dir}{Directory of the data commons projects, as created by \code{\link{init_datacommons}}.}

\item{clone_method}{Means of cloning new repositories; either \code{"http"} (default) or \code{"ssh"}.}

\item{include_distributions}{Logical; if \code{TRUE}, will attempt to locate and cache copies of datasets
pointed to from the data repositories (so far just from Dataverse, implicitly from DOI files).}

\item{only_new}{Logical; if \code{TRUE}, only repositories that do not yet exist will be processed.}

\item{quiet}{Logical; if \code{TRUE}, will not show updated repositories.}
}
\value{
An invisible character vector of updated repositories.
}
\description{
Clone and/or pull repositories that are part of a data commons project.
}
\examples{
\dontrun{
# refresh from a data commons working directory
datacommons_refresh(".")
}
}
