% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{.get_questions}
\alias{.get_questions}
\title{Get Rstudio Community Question Threads}
\usage{
.get_questions()
}
\value{
A tidy tibble of question information.
}
\description{
Get Rstudio Community Question Threads
}
\keyword{internal}
