% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_census_shapes.R
\name{download_census_shapes}
\alias{download_census_shapes}
\title{Download and Prepare Map Shapes}
\usage{
download_census_shapes(dir = NULL, fips = "us", entity = "state",
  name = NULL, year = 2019, resolution = "500k",
  strip_features = FALSE, keep = 0.05, ..., force = FALSE)
}
\arguments{
\item{dir}{Path to a directory in which to save the reformatted files. If these are to be
used in a published site, they should be in the \code{"docs"} directory of the site directory. If not specified,
the shapes are downloaded and returned, but not saved.}

\item{fips}{State Federal Information Processing Standard (FIPS) code, United States Postal Service (USPS)
code, or name; see \href{https://www2.census.gov/geo/docs/reference/state.txt}{state.txt}.}

\item{entity}{Entity name (e.g., \code{"county"}, \code{"tract"}, or \code{"bg"}); see
\href{https://www2.census.gov/geo/tiger/GENZ2019/2019_file_name_def.pdf}{2019_file_name_def.pdf}
(or appropriate year's file).}

\item{name}{Name for the GeoJSON file (without extension) to be written.}

\item{year}{Year of the shapes, 2010 and after.}

\item{resolution}{Resolution of the shapes; one of \code{"500k"} (default), \code{"5m"}, or \code{"20m"}.}

\item{strip_features}{Logical; if \code{TRUE}, will strip all features other than IDs.}

\item{keep}{Proportion of polygon points to retain when simplifying; set to 1 to train all points.
Will not apply by default when there are not more than 50 polygons; set manually to force.}

\item{...}{passes additional arguments to \code{\link[rmapshaper]{ms_simplify}}.}

\item{force}{Logical; if \code{TRUE}, will force a re-download, and overwrite any existing files.}
}
\value{
An \code{sf} \code{data.frame} with a geometry column containing shapes, and other features in columns.
}
\description{
Download cartographic boundary files from the \href{https://www.census.gov}{U.S. Census Bureau},
and reformat them to simplified GeoJSON files.
}
\examples{
\dontrun{
# download Virginia counties shapes
shapes <- download_census_shapes("docs/shapes", "va", "county", name = "counties")
}
}
