% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataverse_data.R,
%   R/download_dataverse_info.R
\name{download_dataverse_data}
\alias{download_dataverse_data}
\alias{download_dataverse_info}
\title{Download Dataverse Dataset Files}
\usage{
download_dataverse_data(id, outdir = tempdir(), files = NULL,
  version = ":latest", server = NULL, load = TRUE, refresh = FALSE,
  verbose = FALSE)

download_dataverse_info(id, server = NULL, refresh = FALSE, verbose = FALSE)
}
\arguments{
\item{id}{The dataset's persistent ID (e.g., \code{"doi:10.18130/V3/SWK71R"}), or a GitHub repository
(\code{"username/repo"}) with a \code{data/dataset_doi} file containing a DOI in its first line.}

\item{outdir}{Directory in which to save files; defaults to a temporary directory.}

\item{files}{Names (full or partial) of files to download, or a number or vector of numbers
identifying file by index as listed on Dataverse; downloads all files in a dataset if not specified.}

\item{version}{Version of the dataset to download. Specifying this will download all files,
even if only a selection is requested.}

\item{server}{Dataverse server; tries to get this from the DOI redirect, but falls back on
\code{Sys.getenv("DATAVERSE_SERVER")}, then \code{getOption("dataverse.server")}, then
\code{"dataverse.lib.virginia.edu"}.}

\item{load}{Logical; if \code{FALSE}, files will be downloaded but not loaded.}

\item{refresh}{Logical; if \code{TRUE}, downloads and replaces any existing files.}

\item{verbose}{Logical; if \code{TRUE}, prints status updates and warnings.}
}
\value{
\code{download_dataverse_data}: If \code{load} is \code{TRUE}, an invisible list with
an entry for each data file if there are multiple files, or the loaded data file if a single file
is requested. Tabular data files are loaded as \code{data.table}s. If \code{load} is \code{FALSE},
a list with the dataset's metadata.

\code{download_dataverse_info}: A list with the dataset's metadata.
}
\description{
Download one or all files in a Dataverse dataset.
}
\examples{
\dontrun{
metadata <- download_dataverse_info("doi:10.18130/V3/SWK71R", verbose = TRUE)
data <- download_dataverse_data("doi:10.18130/V3/SWK71R", verbose = TRUE)
}
}
