% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ct_g_allSamples.R
\name{e_ct_g_allSamples}
\alias{e_ct_g_allSamples}
\title{e_ct_g_allSamples}
\usage{
e_ct_g_allSamples(
  counts,
  anno_cells,
  anno_interactions_allSamples,
  threshold_expr = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{counts}{numeric dataframe \link{0,Inf}: normalized expression data frame containing ligands and receptors in the rows and cells in the columns. Row names and column names should be defined.}

\item{anno_cells}{dataframe: concatenated data frame of the cell annotation from all samples (rows are cell IDs, columns annotation columns: must contain "cell_ID", "cell_type" and "sample_ID" columns).}

\item{anno_interactions_allSamples}{list containing the interactions for each sample.}

\item{threshold_expr}{numeric: expression threshold below which the expression of the ligand (receptor) is not considered. The default value is 0.05.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
anno_interactions_allSamples: list containing the interactions for each sample, with additional columns e_s_l and e_r_r indicating the mean expression in the active cell fraction for the ligand and receptor in the sending and receiving cell types respectively.
}
\description{
The function calculates the mean expression of each gene in all cell types for all samples and a given threshold of expression. It also checks if there are any missing cell types in the annotation of interactions and gives a warning if there are.
}
