% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_params_annoEdges.R
\name{extract_params_annointeractions}
\alias{extract_params_annointeractions}
\title{extract_params_annointeractions}
\usage{
extract_params_annointeractions(anno_interactions, param)
}
\arguments{
\item{anno_interactions}{a list of dataframes, where each dataframe contains the annotation of interactions for a specific sample. Each dataframe should have a column named 'interaction_ID' that serves as a unique identifier for the interactions, and the column of the parameter of interest.}

\item{param}{string, name of the column in the anno_interactions dataframes, containing the parameter of interest.}
}
\value{
dataframe where the rows are interactions and the columns are samples. The values in the dataframe are the values of the parameter of interest in each sample.
}
\description{
collects the values of a specific parameter from multiple dataframes of interactions annotation into one single dataframe.
}
\examples{
# load example data
data('anno_interactions_allSamples')
# extract interactions weights
extract_params_annointeractions(anno_interactions = anno_interactions_allSamples
                       ,param = 'weight'
)

}
