% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_cellType.R
\name{f_cellType}
\alias{f_cellType}
\title{f_cellType}
\usage{
f_cellType(
  anno_interactions,
  cell_type,
  anno_cells,
  threshold_celltype_size = 4
)
}
\arguments{
\item{anno_interactions}{dataframe containing interactions annotation for a given sample.}

\item{cell_type}{string : the cell type of interest.}

\item{anno_cells}{dataframe: concatenated data frame of the cell annotation from all samples (rows are cell IDs, columns annotation columns: must contain "cell_ID", "cell_type" and "sample_ID" columns).}

\item{threshold_celltype_size}{numeric \link{0,Inf}: threshold of minimum number of cells in a cell type in a sample. The default value is 4.}
}
\value{
f numeric \link{0,1}: fraction of cells of the given cell type in the sample.
}
\description{
calculates the fraction of cells of a given cell type in a sample
}
\examples{
anno_interactions <- data.frame(sending_cell_type = c('A', 'B', 'C', 'D', 'E')
                        ,receiving_cell_type = c('A', 'B', 'C', 'D', 'E')
                       ,sending_cell = c('A1', 'B1', 'C1', 'D1', 'E1')
                      ,receiving_cell = c('A2', 'B2', 'C2', 'D2', 'E2')
                     ,sending_gene = c('A', 'B', 'C', 'D', 'E')
                   ,receiving_gene = c('A', 'B', 'C', 'D', 'E')
                 ,sending_gene_expression = c(1, 2, 3, 4, 5)
              ,receiving_gene_expression = c(1, 2, 3, 4, 5)
           ,sending_cell_expression = c(1, 2, 3, 4, 5)
       ,receiving_cell_expression = c(1, 2, 3, 4, 5)
    ,sending_cell_type_expression = c(1, 2, 3, 4, 5)
,receiving_cell_type_expression = c(1, 2, 3, 4, 5)
)
cell_type <- 'A'
anno_cells <- data.frame(cell_type = c('A', 'B', 'C', 'D', 'E')
                       ,cell = c('A1', 'B1', 'C1', 'D1', 'E1')
)
f_cellType(anno_interactions
          ,cell_type
         ,anno_cells
       ,threshold_celltype_size = 4
)

}
