% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_cellType_max.R
\name{f_cellType_max}
\alias{f_cellType_max}
\title{f_cellType_max}
\usage{
f_cellType_max(
  anno_cells,
  anno_interactions_allSamples,
  threshold_celltype_size = 4,
  verbose = FALSE
)
}
\arguments{
\item{anno_cells}{dataframe: concatenated data frame of the cell annotation from all samples (rows are cell IDs, columns annotation columns: must contain "cell_ID", "cell_type" and "sample_ID" columns).}

\item{anno_interactions_allSamples}{a list of data frames containing annotation information for interactions in all samples.}

\item{threshold_celltype_size}{numeric: threshold of minimum number of cells in a cell type in a sample. The default value is 4.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
anno_interactions_allSamples: list of dataframes (numeric \link{0,1}) with additional columns for the maximum cell type abundance of the sending cell type, and of the receiving cell type.
}
\description{
calculates the maximum fraction of each cell type in all samples, and stores it in the anno_interactions_allSamples list.
}
\examples{
# calculate the maximum fraction of each cell type in all samples
anno_interactions_allSamples <- f_cellType_max(anno_cells
                                          ,anno_interactions_allSamples
                                        ,verbose = TRUE
                                     ,threshold_celltype_size = 4
                                  )

}
