% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_ct_allSamples.R
\name{f_ct_allSamples}
\alias{f_ct_allSamples}
\title{f_ct_allSamples}
\usage{
f_ct_allSamples(
  anno_cells,
  anno_interactions_allSamples,
  threshold_celltype_size = 4,
  verbose = FALSE
)
}
\arguments{
\item{anno_cells}{dataframe: concatenated data frame of the cell annotation from all samples (rows are cell IDs, columns annotation columns: must contain "cell_ID", "cell_type" and "sample_ID" columns).}

\item{anno_interactions_allSamples}{a list of data frames containing annotation information for interactions in all samples.}

\item{threshold_celltype_size}{numeric \link{0,Inf}: threshold of minimum number of cells in a cell type in a sample. The default value is 4.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
a dataframe (numeric \link{0,1}) with fractions of cell types where columns are samples and rows are cell types.
}
\description{
calculates fraction of each cell type in each sample.
}
\examples{
# calculate fraction of each cell type in each sample
anno_interactions_allSamples <- f_ct_allSamples(anno_cells
                                        ,anno_interactions_allSamples
                                       ,threshold_celltype_size = 4
                                      ,verbose = TRUE
)

}
