% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDB.R
\name{filter_pairs_with_ppi}
\alias{filter_pairs_with_ppi}
\title{Filter Pairwise Pairs Based on PPI Network}
\usage{
filter_pairs_with_ppi(pairwise_pairs)
}
\arguments{
\item{pairwise_pairs}{A dataframe containing pairwise pairs of ligand-receptor interactions.}
}
\value{
A dataframe of filtered pairwise pairs that exist in the PPI network.
}
\description{
Filters the pairwise pairs of ligand-receptor interactions by checking against PPI network.
}
\examples{
both_db <- import_db("both")
pairwise_pairs <- create_pairwise_pairs(both_db)
filtered_pairs <- filter_pairs_with_ppi(pairwise_pairs)
}
