% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_stat.R
\name{general_stat}
\alias{general_stat}
\title{general_stat}
\usage{
general_stat(comm_result, verbose = FALSE)
}
\arguments{
\item{comm_result:}{list containing information about the interactions, samples, and their annotations.}

\item{verbose:}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
comm_results: The dataframe anno_interactions gets following new columns:
\itemize{
\item log2FC_weights
\item log2FC_phi_l_s
\item log2FC_phi_r_r
\item log2FC_phi
\item log2FC_p_l_s
\item log2FC_p_r_r
\item log2FC_p
}
}
\description{
calculates the log2 fold change between the mean values in the case cohort and the mean values in the control cohort for the following parameters: weights, rho_s, rho_r, rho, phi_l_s, phi_r_r, phi, p_l_s, p_r_r, and p.
}
\examples{
# load example data
data('comm_result')
# calculate general statistics
comm_result <- general_stat(comm_result)
}
