% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_datacommons.R
\name{init_datacommons}
\alias{init_datacommons}
\title{Initialize a Data Commons}
\usage{
init_datacommons(dir, name = "Data Commons", repos = NULL,
  default_user = "", remote = NULL, url = NULL, refresh_after = FALSE,
  overwrite = FALSE, serve = FALSE, host = "127.0.0.1", port = 3000,
  use_local = FALSE, verbose = interactive())
}
\arguments{
\item{dir}{Path to the desired data commons directory.}

\item{name}{Name of the data commons.}

\item{repos}{A vector of repository names to add to \code{commons.json}.}

\item{default_user}{GitHub username to prepend to repository names if needed.}

\item{remote}{Name of the data commons' GitHub repository (\code{"username/reponame"}).}

\item{url}{URL of the data commons' monitor site; defaults to the GitHub Pages URL associated with \code{remote}
if provided (\code{"https://username.github.io/reponame"}).}

\item{refresh_after}{Logical; if \code{FALSE}, will not run \code{\link{datacommons_refresh}}
after initiating the project. Defaults to \code{TRUE} when first creating a data commons project.}

\item{overwrite}{Logical; if \code{TRUE}, will overwrite existing datacommons files in \code{dir}.
The included \code{.js} and \code{.sh} files are always rewritten, and if \code{name},
\code{repos}, or \code{default_user} is specified, \code{commons.json} will also be rewritten
regardless of \code{overwrite}.}

\item{serve}{Logical; if \code{TRUE}, will serve the \code{docs} directory.}

\item{host}{The IPv4 address to listen to if \code{serve} is \code{TRUE}; defaults to \code{"127.0.0.1"}.}

\item{port}{The port to listen on if \code{serve} is \code{TRUE}; defaults to 3000.}

\item{use_local}{Logical; if \code{TRUE}, will use a \code{datacommons.js} script located in
a local \code{dist/docs/dev} directory, relative to \code{dir}.}

\item{verbose}{Logical; if \code{FALSE}, suppresses messages.}
}
\value{
Path to the datacommons directory.
}
\description{
Initialize a project to keep track of separate dataset repositories and distributions.
}
\details{
The shell scripts included in the project's \code{scripts} directory can be used to retrieve
and update repositories over SSH.

This will clone or pull repositories listed in \code{scripts/repos.txt}:
\code{sh scripts/get_repos.sh}

This will add, commit, and push all changes in all repositories:
\code{sh scripts/update_repos.sh "commit message"}
}
\examples{
\dontrun{
init_datacommons(
  "../datacommons",
  name = "Data Commons",
  remote = ""
)
}
}
