% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_function.R
\name{init_function}
\alias{init_function}
\title{Create a new package function}
\usage{
init_function(name, dir = ".", overwrite = FALSE)
}
\arguments{
\item{name}{Name of the function. Should start with the function's category, followed by a specific name,
separated by an underscore (\code{'_'}) (e.g., \code{'init_function'}).}

\item{dir}{Path to the package's development directory; default is the current working directory.}

\item{overwrite}{Logical; if \code{TRUE}, replaces existing files with templates.}
}
\value{
Creates files in \code{dir/R} and \code{dir/tests/testthat}, attempts to navigate to the code file,
and returns a character vector to their paths.
}
\description{
Create initial script and test files for a function that is to be added to a package.
}
\examples{
\dontrun{

# creates a skeleton for a `measure_new` function
init_function("measure_new")
}
}
