% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_repository.R
\name{init_repository}
\alias{init_repository}
\title{Create a Data Repository}
\usage{
init_repository(dir, datasets = "dataset", init_data = TRUE,
  init_site = TRUE, init_git = TRUE, overwrite = FALSE,
  quiet = !interactive())
}
\arguments{
\item{dir}{Directory in which to create the repository's structure. Will be created if it does not exist.}

\item{datasets}{A character vector of dataset names; for each of these, a subdirectory will be made
containing \code{code} and \code{data} directories.}

\item{init_data}{Logical; if \code{FALSE}, will not run \code{\link{init_data}} on the repository.}

\item{init_site}{Logical; if \code{FALSE}, will not run \code{\link{init_site}} on the repository.}

\item{init_git}{Logical; if \code{FALSE}, will not run \code{git init} on the repository.}

\item{overwrite}{Logical; if \code{TRUE}, will overwrite existing site files in \code{dir}.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
Path to the created repository directory.
}
\description{
Create a repository for a dataset, which may include data documentation and/or a data site.
}
\examples{
\dontrun{
# initialize repository in the current working directory
init_repository(".")
}
}
