% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_shiny.R
\name{init_shiny}
\alias{init_shiny}
\title{Create a new Shiny app}
\usage{
init_shiny(name, dir = ".", overwrite = FALSE)
}
\arguments{
\item{name}{Name of the app, corresponding to the directory name.}

\item{dir}{Directory in which to make the parent directory of the app.}

\item{overwrite}{Logical; if \code{TRUE}, existing files will be overwritten.}
}
\value{
Path to the created app directory.
}
\description{
Create the basic structure of a Shiny application.
}
\examples{
\dontrun{
init_shiny("test")
}
}
