% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_site.R
\name{init_site}
\alias{init_site}
\title{Create a website}
\usage{
init_site(title = "app", dir = ".", with_data = TRUE, overwrite = FALSE)
}
\arguments{
\item{title}{Title of the site.}

\item{dir}{Directory in which to create the site's directory.}

\item{with_data}{logical; if \code{FALSE}, a data sub-directory and package will not be created.}

\item{overwrite}{logical; if \code{TRUE}, will overwrite existing site files in \code{dir}.}
}
\value{
Path to the created site directory.
}
\description{
Create a repository for a static, website for data documentation and exploration.
}
\examples{
\dontrun{
init_site("app")
}
}
