% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_buttongroup.R
\name{input_buttongroup}
\alias{input_buttongroup}
\title{Add a button toggle group to a website}
\usage{
input_buttongroup(label, options, default = 0, display = options,
  id = label, button_class = NULL, variable = NULL, dataset = NULL,
  depends = NULL)
}
\arguments{
\item{label}{Label of the input for the user.}

\item{options}{A vector of options, the name of a variable from which to pull levels, or either \code{"datasets"}
or \code{"variables"} to select names of datasets or variables.}

\item{default}{Which of the options to default to; either its index or value.}

\item{display}{A display version of the options.}

\item{id}{Unique id of the element to be created.}

\item{button_class}{Class name to add to each label (which appears as the button).}

\item{variable}{The name of a variable from which to get levels (overwritten by \code{depends}).}

\item{dataset}{The name of an included dataset, where \code{variable} should be looked for; only applies when
there are multiple datasets with the same variable name.}

\item{depends}{The id of another input on which the options depend; this will take president over \code{dataset}
and \code{variable}, depending on this type of input \code{depends} points to.}
}
\value{
A character vector of the contents to be added.
}
\description{
Adds a group of radio-like buttons, where only one can be selected at a time.
}
\examples{
\dontrun{
input_buttongroup()
}
}
