% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checkbox.R
\name{input_checkbox}
\alias{input_checkbox}
\title{Add checkboxes, radio buttons, or switches to a website}
\usage{
input_checkbox(label, options, default = "all", display = options,
  id = label, variable = NULL, dataset = NULL, depends = NULL,
  multi = TRUE, as.switch = FALSE)
}
\arguments{
\item{label}{Label of the input for the user.}

\item{options}{A vector of options, the name of a variable from which to pull levels, or either \code{"datasets"}
or \code{"variables"} to select names of datasets or variables.}

\item{default}{A vector of items to check by default (or "all" or "none") if \code{multi} is \code{TRUE}, or
the option to select by default.}

\item{display}{A display version of the options.}

\item{id}{Unique id of the element to be created.}

\item{variable}{The name of a variable from which to get levels (overwritten by \code{depends}).}

\item{dataset}{The name of an included dataset, where \code{variable} should be looked for; only applies when
there are multiple datasets with the same variable name.}

\item{depends}{The id of another input on which the options depend; this will take president over \code{dataset}
and \code{variable}, depending on this type of input \code{depends} points to.}

\item{multi}{Logical; if \code{FALSE}, only one option can be selected at a time, turning the checkboxes into radio
buttons.}

\item{as.switch}{Logical; if \code{TRUE}, displays checkboxes or radio buttons as switches.}
}
\value{
A character vector of the contents to be added.
}
\description{
Adds a set of checkbox, radio buttons, or switches to a website.
}
\examples{
\dontrun{
input_checkbox("Label", c("a", "b", "c"))
}
}
