% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_dataview.R
\name{input_dataview}
\alias{input_dataview}
\title{Add a unified data view to a website}
\usage{
input_dataview(dataset = NULL, ids = NULL, features = NULL, id = NULL)
}
\arguments{
\item{dataset}{Select which dataset to pull from; the name of an included dataset, or ID of a
selector of dataset names.}

\item{ids}{Select which IDs to include; a vector of IDs that appear in the specified dataset, or the ID of a
selector of IDs. If an IDs map is included in the site's datapackage, mapped components can be referred to here.}

\item{features}{Select IDs based on their features; a named list or vector, with names corresponding to
the names of features included in an \code{ids} field of the site's datapackage, and values corresponding to
a value or vector of values, or a selector of values.}

\item{id}{An ID of the data view, for association with output elements. Defaults to \code{'view'} appended
with a zero-index of views, based on the order in which they were specified
(e.g., \code{'view0'} for the first view).}
}
\value{
The entered filter.
}
\description{
Collects specified inputs to create a filtered dataset from which outputs can pull.
}
\examples{
\dontrun{
input_dataview(list())
}
}
