% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_number.R
\name{input_number}
\alias{input_number}
\title{Add a number input to a website}
\usage{
input_number(label, id = label, ..., default = NULL, variable = NULL,
  min = NULL, max = NULL, step = NULL, type = "number", class = NULL,
  dataview = NULL, floating_label = TRUE)
}
\arguments{
\item{label}{Label of the input for the user.}

\item{id}{Unique ID of the element to be created.}

\item{...}{Other attributes to add to the input.}

\item{default}{Default value of the input, the ID of an input to use as the default, or \code{"min"} or \code{"max"}
to default to the current minimum or maximum value.}

\item{variable}{The name of a variable or ID of a variable selector to get a range from.}

\item{min}{The smallest allowed value.}

\item{max}{The largest allowed value.}

\item{step}{Amount to increase or decrease the value by when changed with arrows.}

\item{type}{Name of the input's type -- other number-based types like \code{date} might make sense.}

\item{class}{Class names to add to the input's list.}

\item{dataview}{ID of a \code{\link{input_dataview}}, to use as a source of variables.}

\item{floating_label}{Logical; if \code{FALSE}, labels are separate from their input elements.}
}
\value{
A character vector of the contents to be added.
}
\description{
Adds an direct number input element to a website.
}
\examples{
\dontrun{
input_text("entered_text", "Enter Text:")
}
}
