% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_rule.R
\name{input_rule}
\alias{input_rule}
\title{Add an input rule to a website}
\usage{
input_rule(condition, effects)
}
\arguments{
\item{condition}{A string representing the condition (e.g., "input_a == 'a'"), with multiple conditions
separated by \code{"&"}. If all conditions are \code{TRUE}, all \code{effects} will be set.}

\item{effects}{A list with names corresponding to input IDs, and values of what they should be set to.}
}
\value{
The entered condition and effects.
}
\description{
Specifies if-then conditions for inputs; that is, when one input is changed, change another based on
the entered conditions.
}
\examples{
\dontrun{
input_select("input_a", c("a", "b", "c"))
input_slider("input_b", c(0, 10))
input_rule("input_a == 'a' && input_b != 0", list(input_b = 10))
}
}
