% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_select.R
\name{input_select}
\alias{input_select}
\title{Add a select input to a website}
\usage{
input_select(label, options, default = -1, display = options, id = label,
  ..., note = NULL, group_feature = NULL, variable = NULL,
  dataset = NULL, depends = NULL, dataview = NULL, subset = "all",
  filters = NULL, reset_button = FALSE, button_class = NULL,
  as.row = FALSE, floating_label = TRUE)
}
\arguments{
\item{label}{Label of the input for the user.}

\item{options}{A vector of options, the name of a variable from which to pull levels, or \code{"datasets"},
\code{"variables"}, \code{"ids"}, or \code{"palettes"} to select names of datasets, variables, entity ids, or
color palettes.}

\item{default}{Which of the options to default to; either its index or value.}

\item{display}{A display version of the options.}

\item{id}{Unique ID of the element to be created.}

\item{...}{Additional attributes to set on the select element.}

\item{note}{Text to display as a tooltip for the input.}

\item{group_feature}{Name of a measure or entity feature to use as a source of option grouping,
if \code{options} is \code{"variables"} or \code{"ids"}.}

\item{variable}{The name of a variable from which to get levels (overwritten by \code{depends}).}

\item{dataset}{The name of an included dataset, where \code{variable} should be looked for; only applies when
there are multiple datasets with the same variable name.}

\item{depends}{The ID of another input on which the options depend; this will take president over \code{dataset}
and \code{variable}, depending on this type of input \code{depends} points to.}

\item{dataview}{The ID of an \code{\link{input_dataview}}, used to filter the set of options, and potentially
specify dataset if none is specified here.}

\item{subset}{Determines the subset of options shown if \code{options} is \code{"ids"}; mainly \code{"all"}
(default) to apply all filters, including the current selection, or \code{"full_filter"} to apply all
feature and variable filters, but not the current selection.}

\item{filters}{A list with names of \code{meta} entries (from \code{variable} entry in \code{\link{data_add}}'s
\code{meta} list), and values of target values for those entries, or the IDs of value selectors.}

\item{reset_button}{If specified, adds a button after the select element that will revert the selection
to its default; either \code{TRUE}, or text for the reset button's label.}

\item{button_class}{Class name to add to the reset button.}

\item{as.row}{Logical; if \code{TRUE}, the label and input are in separate columns within a row.}

\item{floating_label}{Logical; if \code{FALSE} or \code{as.row} is \code{TRUE}, labels are separate from
their inputs.}
}
\value{
A character vector of the contents to be added.
}
\description{
Adds an input to select from the entered options.
}
\examples{
\dontrun{
input_select()
}
}
