% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_slider.R
\name{input_slider}
\alias{input_slider}
\title{Add a range slider to a website}
\usage{
input_slider(label, range = c(0, 1), stepsize = 1, default = NULL,
  id = label, variable = NULL, dataset = NULL, depends = NULL)
}
\arguments{
\item{label}{Label of the input for the user.}

\item{range}{A numeric vector with a minimum and maximum, or the name of a variable to calculate a range from.}

\item{stepsize}{How much moving the handle adjusts the selected value.}

\item{default}{Starting value of the slider handle.}

\item{id}{Unique id of the element to be created.}

\item{variable}{The name of a variable to calculate a range from (overwrites \code{range}).}

\item{dataset}{The name of an included dataset, where \code{variable} should be looked for; only applies when
there are multiple datasets with the same variable name.}

\item{depends}{The id of another input on which the options depend; this will take president over \code{dataset}
and \code{variable}, depending on this type of input \code{depends} points to.}
}
\value{
A character vector of the contents to be added.
}
\description{
Adds an input to select within the entered range.
}
\examples{
\dontrun{
input_slider()
}
}
