% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_switch.R
\name{input_switch}
\alias{input_switch}
\title{Add a single switch or checkbox to a website}
\usage{
input_switch(label, id = label, ..., default_on = TRUE, as.checkbox = FALSE)
}
\arguments{
\item{label}{Label of the input for the user.}

\item{id}{Unique id of the element to be created.}

\item{...}{Additional attributes to set on the element.}

\item{default_on}{Logical; if \code{FALSE}, the switch will start off.}

\item{as.checkbox}{Logical; if \code{TRUE}, display the switch as a checkbox.}
}
\value{
A character vector of the contents to be added.
}
\description{
Adds a single toggle, displayed as a switch or checkbox to a website.
}
\examples{
\dontrun{
input_switch("Label")
}
}
\seealso{
For a group of switches, checkboxes, or radio buttons, use \code{\link{input_checkbox}}.
}
