% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_variable.R
\name{input_variable}
\alias{input_variable}
\title{Add an internal variable to a website}
\usage{
input_variable(id, cases, default = "")
}
\arguments{
\item{id}{Name of the variable, which can be referred to by other inputs of outputs.}

\item{cases}{A list of conditions with names specifying conditions, and values to set the variable to
in that condition (e.g., \code{list("input_a == 1" = 1)}).
These can also be specified separately with an \code{\link{input_rule}}.}

\item{default}{The value to set if no condition is \code{TRUE}.}
}
\value{
A version of the resulting variable object.
}
\description{
Creates an internal variable (virtual input), which can be used for display logic.
}
\examples{
\dontrun{
input_select("input_a", c("a", "b", "c"))
input_variable("vinput_a", list("input_a == c" = "b"), "a")

# vinput_a will be "a" unless input_a is "c"
}
}
