% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_classes.R
\name{interaction_classes}
\alias{interaction_classes}
\title{interaction_classes}
\usage{
interaction_classes(my_interactions, threshold_log2FC_components = 1)
}
\arguments{
\item{my_interactions}{list of interactions (see \code{\link{interaction_list}})}

\item{threshold_log2FC_components}{threshold for the log2FC of the components}
}
\value{
my_interactions list of interactions (see \code{\link{interaction_list}})
}
\description{
This function adds columns to the annotation of the interactions
}
\examples{
# load the data
data("my_interactions")
# add the columns
my_interactions <- interaction_classes(my_interactions
                                     ,threshold_log2FC_components = 1)
# show the first 10 rows
head(my_interactions$anno_interactions
   ,n = 10)
# show the last 10 rows
tail(my_interactions$anno_interactions
  ,n = 10)

}
