% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2FC.R
\name{log2FC}
\alias{log2FC}
\title{log2FC}
\usage{
log2FC(x, anno_samples, verbose = FALSE)
}
\arguments{
\item{x}{numeric dataframe: can be a dataframe with interactions weights, phi, p values, etc. The rows are interaction_IDs, the columns are sample IDs. The column names should be identical to the row names in anno_samples.}

\item{anno_samples}{dataframe: should contain the columns: sample_ID and case_or_control. The row names should be defined and equal to sample_ID.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
numeric \link{-Inf, Inf}: log2FC for each value of the submitted parameter.
}
\description{
calculates log2FC for a given parameter between the case and the control level. To avoid division by zero, all zero values are substituted by a pseudo-count. The pseudo-count is calculated as the minimum value in the x divided by 10.
}
\examples{
# load example data
data("comm_result")
# calculate general statistics
comm_result <- general_stat(comm_result)
# calculate log2FC for interactions weights
log2FC(x = comm_result$weights
      ,anno_samples = comm_result$anno_samples
     ,verbose = TRUE
)
# calculate log2FC for phi
log2FC(x = comm_result$phi
     ,anno_samples = comm_result$anno_samples
    ,verbose = TRUE
)
}
