% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_database.R
\docType{data}
\name{LR_database}
\alias{LR_database}
\title{Ligand Receptor Database}
\format{
Ligand Receptor Database, each row is a ligand-receptor pair.
}
\usage{
data(LR_database)
}
\description{
Data from previous studies of the ligand-receptor interaction.
}
\examples{
data(LR_database)
head(LR_database)
}
\references{
OmniPath: guidelines and gateway for literature-curated signaling pathway resources
}
\keyword{datasets}
