% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDB.R
\name{map_gene_data}
\alias{map_gene_data}
\title{Map Gene Symbols to Descriptions in Ligand-Receptor Interactions}
\usage{
map_gene_data(complete)
}
\arguments{
\item{complete}{A data frame containing ligand-receptor interactions with columns 'Ligand' and 'Receptor'.}
}
\value{
The input data frame with additional columns 'Ligand.Name' and 'Receptor.Name' containing the mapped protein descriptions. The function also adjusts specific gene symbols and reorders the columns.
}
\description{
This function maps gene symbols to their corresponding descriptions in a complete dataset of ligand-receptor interactions. It utilizes mygene library for protein descriptions.
}
\examples{
# Assuming 'complete' is a data frame with ligand-receptor interactions
complete_mapped <- map_gene_data(complete)

}
