% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_e_ct_g_condition.R
\name{mean_e_ct_g_condition}
\alias{mean_e_ct_g_condition}
\title{mean_e_ct_g_condition}
\usage{
mean_e_ct_g_condition(comm_result)
}
\arguments{
\item{comm_result:}{list containing information about the interactions, samples, and their annotations.}
}
\value{
comm_result: The dataframe anno_interactions gets following new columns:
\itemize{
\item mean_e_s_l_control,
\item mean_e_s_l_case,
\item mean_e_r_r_control,
\item mean_e_r_r_case,
}
}
\description{
calculates the mean for the parameters e_s_l and e_r_r for the case and control cohorts.
}
\examples{
# load example data
data('comm_result')
# calculate mean_e_ct_g_condition
comm_result <- mean_e_ct_g_condition(comm_result)

}
