% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_to_max.R
\name{norm_to_max}
\alias{norm_to_max}
\title{norm_to_max}
\usage{
norm_to_max(param, param_max)
}
\arguments{
\item{param}{a numeric vector representing a parameter of interest for normalization.}

\item{param_max}{a numeric vector representing the maximum value of the parameter.}
}
\value{
param_norm a numeric vector of normalized parameter values.
}
\description{
normalizes the param for the cell type of interest to param_max.
}
