% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_interactions_for_forest.R
\name{order_interactions_for_forests}
\alias{order_interactions_for_forests}
\title{order_interactions_for_forests}
\usage{
order_interactions_for_forests(my_idx)
}
\arguments{
\item{my_idx}{A vector of indices for subsetting the interactions dataframe}
}
\value{
A sorted dataframe of interactions
}
\description{
This function takes in a vector of indices and subsets the interactions dataframe based on the indices. Then it sorts the interactions based on log2FC_weights, log2FC_p_s_l, log2FC_phi_s_l, log2FC_rho_s, log2FC_p_r_r, log2FC_phi_r_r, log2FC_rho_r and interaction_ID columns. It also has helper functions like cluster_interactions and pick_param which are used in the sorting process.
}
\examples{
# load example data
data("comm_result")
# calculate general statistics
comm_result <- general_stat(comm_result)
# order interactions
order_interactions_for_forests(my_idx = 1:10)

}
