% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_info.R
\name{output_info}
\alias{output_info}
\title{Add a tooltip-like display to a website}
\usage{
output_info(title = NULL, body = NULL, row_style = "table",
  default = NULL, dataview = NULL, variable = NULL, subto = NULL,
  id = NULL, variable_info = TRUE, floating = FALSE)
}
\arguments{
\item{title}{Title text, or the source of title text.}

\item{body}{A list of entries in the info body section, which can be raw text or references to features or
data variables.}

\item{row_style}{A character specifying how rows should be displayed: \code{"table"} (default)
to place names and values in separate columns of a table row, or \code{"stack"} to place names over values.
Repeats over rows.}

\item{default}{A list with entries for \code{"title"} and \code{"body"}, which are treated as raw text.}

\item{dataview}{The ID of a dataview, used for the persistent display.}

\item{variable}{Name of the variable from which to display variable information and values. If not specified,
this will default to the coloring variable of maps and plots, or the y variable of a dataview.}

\item{subto}{A vector of output IDs to receive hover events from.}

\item{id}{Unique id of the element.}

\item{variable_info}{Logical; if \code{TRUE} (default), variable names can be clicked for more information.}

\item{floating}{Logical; if \code{TRUE}, the information pane will appear next to the cursor.}
}
\value{
A character vector of the content to be added.
}
\description{
Adds an output to display information about a hovered-over or selected data point.
}
\examples{
\dontrun{
output_info("Initial View", "Hover over plot elements for more information.")
}
}
