% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_map.R
\name{output_map}
\alias{output_map}
\title{Add a map to a webpage}
\usage{
output_map(shapes = NULL, color = NULL, color_time = NULL,
  dataview = NULL, legend = TRUE, click = NULL, subto = NULL,
  options = list(), tiles = list(url =
  "https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png", options = list(maxZoom
  = 19)))
}
\arguments{
\item{shapes}{A named list or vector with URLs of GeoJSON files.}

\item{color}{The name of a variable, or id of a variable selector, to be used to color polygons.}

\item{color_time}{The ID of a selector to specify which timepoint of \code{color} to use.}

\item{dataview}{The ID of an \code{\link{input_dataview}} component.}

\item{legend}{Logical; if \code{FALSE}, the legend will not be shown.}

\item{click}{The ID of an input to set to a clicked polygon's ID.}

\item{subto}{A vector of output IDs to receive hover events from.}

\item{options}{A list of configuration options, potentially extracted from a saved leaflet object (see
\href{https://leafletjs.com/reference-1.7.1.html#map-example}{Leaflet documentation}).}

\item{tiles}{A list or list of lists containing provider information (see
\href{https://leaflet-extras.github.io/leaflet-providers/preview/}{leaflet providers}; e.g.,
\code{list(}\code{url = "https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"}, \code{options = list(maxZoom = 19))}).}
}
\value{
A character vector of the content to be added.
}
\description{
Adds a Leaflet map to a webpage, based on specified or selected inputs.
}
\examples{
\dontrun{
output_map()
}
}
