% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_plot.R
\name{output_plot}
\alias{output_plot}
\title{Add a plot to a webpage}
\usage{
output_plot(x = NULL, y = NULL, color = NULL, color_time = NULL,
  dataview = NULL, click = NULL, subto = NULL, options = list())
}
\arguments{
\item{x}{The name of a variable to plot along the x-axis.}

\item{y}{The name of a variable to plot along the y-axis.}

\item{color}{The name of a variable, or ID of a variable selector to use to color lines.}

\item{color_time}{The ID of a selector to specify which timepoint of \code{color} to use.}

\item{dataview}{The ID of an \code{\link{input_dataview}} component.}

\item{click}{The ID of an input to set to a clicked line's ID.}

\item{subto}{A vector of output IDs to receive hover events from.}

\item{options}{A list of configuration options, with named entries for any of \code{data}, \code{layout},
or \code{options}, potentially extracted from a saved plotly object (see
\href{https://plotly.com/javascript/configuration-options}{Plotly documentation}).}
}
\value{
A character vector of the content to be added.
}
\description{
Adds a Plotly plot to a webpage, based on specified or selected variables.
}
\examples{
\dontrun{
# for mpg ~ wt * am in a site based on mtcars data
output_plot("wt", "mpg", "am")
}
}
