% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_table.R
\name{output_table}
\alias{output_table}
\title{Add a table to a webpage}
\usage{
output_table(variables = NULL, dataset = NULL, dataview = NULL,
  options = NULL, class = "compact")
}
\arguments{
\item{variables}{Variable names in the included data, or the ID of a
variable selector to show in columns of the table.}

\item{dataset}{The name of a dataset, or ID of a dataset selector, to find \code{variables} in;
used if \code{dataview} is not specified.}

\item{dataview}{The ID of an \code{\link{input_dataview}} component.}

\item{options}{A list of configuration options, see
\href{https://datatables.net/reference/option}{DataTables Documentation}.}

\item{class}{Class names to add to the table.}
}
\value{
A character vector of the content to be added.
}
\description{
Adds a DataTables table to a webpage, based on specified or selected .
}
\examples{
\dontrun{
output_table()
}
}
