% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_text.R
\name{output_text}
\alias{output_text}
\title{Add dynamic text to a website}
\usage{
output_text(text, tag = "p", id = NULL, class = NULL)
}
\arguments{
\item{text}{A vector of text to be parsed.}

\item{tag}{Tag name of the element containing the text.}

\item{id}{Unique ID of the output element.}

\item{class}{Class names to add to the output's element.}
}
\value{
A character vector of the containing element of the text.
}
\description{
Adds a textual output based on the current state of input elements.
}
\examples{
\dontrun{
output_text("Selection: input_a{{input_a}, }")
}
}
