% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{p}
\alias{p}
\title{p}
\usage{
p(p_s_l, p_r_r, verbose = FALSE)
}
\arguments{
\item{p_s_l}{numeric \link{0,1}: probability of interaction of a ligand with its counterpart receptor based on the relative mean expression of this ligand in the sending cell type.}

\item{p_r_r}{numeric \link{0,1}: probability of interaction of a receptor with its counterpart ligand based on the relative mean expression of this receptor in the receiving cell type.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
numeric \link{0,1}: probability of interaction based on the relative mean expression of the ligand and receptor in the sending and receiving cell types.
}
\description{
calculates the expression based probability of communication
}
\examples{
p_s_l <- 0.1
p_r_r <- 0.1
p_expr(p_s_l = p_s_l
     ,p_r_r = p_r_r
    ,verbose = TRUE
)

}
