% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_cellType_gene.R
\name{p_cellType_gene}
\alias{p_cellType_gene}
\title{p_cellType_gene}
\usage{
p_cellType_gene(e_cellType_gene, e_cellType_gene_max, verbose = FALSE)
}
\arguments{
\item{e_cellType_gene}{numeric \link{0,Ifn}: mean expression level of the ligand (or receptor) in the active fraction of the cell type of interest.}

\item{e_cellType_gene_max}{numeric \link{0,Ifn}: maximum expression level of the ligand (or receptor) in the cell type of interest across all samples.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
p_expr numeric \link{0,1}: probability of interaction, normalized to the maximum expression level across all cell types.
}
\description{
calculates the probability of interaction of a ligand (or receptor) with its counterpart receptor (or ligand) based on its expression level in the cell type of interest.
}
\examples{
e_cellType_gene <- 0.1
e_cellType_gene_max <- 0.1
p_cellType_gene(e_cellType_gene = e_cellType_gene
              ,e_cellType_gene_max = e_cellType_gene_max
             ,verbose = TRUE
)

}
