% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_menu.R
\name{page_menu}
\alias{page_menu}
\title{Adds a menu to a website}
\usage{
page_menu(..., position = "right", width = NULL, height = NULL,
  collapsible = TRUE, default_open = FALSE, wraps = TRUE, sizes = NA,
  breakpoints = NA)
}
\arguments{
\item{...}{Elements to appear in the menu.}

\item{position}{A character indicating which side to place the menu on; one of
\code{"top"}, \code{"right"}, \code{"bottom"}, or \code{"left"}.}

\item{width}{Width of the menu, if \code{position} if \code{"left"} or \code{"right"}.}

\item{height}{Height of the menu, if \code{position} if \code{"top"} or \code{"bottom"}.}

\item{collapsible}{Logical; if \code{FALSE}, removes the button to open or close the menu.}

\item{default_open}{Logical; if \code{TRUE}, the menu will start open.}

\item{wraps}{Logical or character (\code{"row" or "col"}); if \code{FALSE},
elements of the menu will not be wrapped in row or column elements.}

\item{sizes}{The relative size of each wrapper, between 1 and 12, or \code{"auto"}; default is equal size.}

\item{breakpoints}{Bootstrap breakpoint of each wrapper; one of \code{""} (extra small), \code{"sm"},}
}
\value{
A character vector of the content to be added.
}
\description{
Adds a section to arrange inputs, that can optionally be collapsed, and positioned on any side.
}
\examples{
\dontrun{
page_menu(
  input_select("menu item", c("a", "b", "c"))
)
}
}
