% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_navbar.R
\name{page_navbar}
\alias{page_navbar}
\title{Add a navbar (header) section to a website}
\usage{
page_navbar(..., title = "", logo = "", breakpoint = "md")
}
\arguments{
\item{...}{Content to add to the navbar. Can be a lists specifying submenus and their items.}

\item{title}{A title to appear in the header.}

\item{logo}{URL of an image to appear in the header.
Each list can have entries for \code{"name"} (the navbar button text), \code{"title"} (the menu title),
\code{"id"} (the ID of the menu element), \code{"class"} (class of the menu element),
\code{"placement"} (start, top, end, or bottom), \code{"backdrop"} (logical), and \code{items}
(a list of menu items).}

\item{breakpoint}{Bootstrap breakpoint code specifying the width at which the menu collapses;
default is \code{"md"}.}
}
\value{
A character vector of the content to be added.
}
\description{
Adds a bar at the top of the page, potentially with a title and logo image.
}
\examples{
page_navbar(
  title = "Site Name",
  list(
    name = "menu",
    title = "Menu",
    items = list(input_select("menu select", c("a", "b")))
  )
)
}
