% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_text.R
\name{page_text}
\alias{page_text}
\title{Add static text elements to a website.}
\usage{
page_text(text, ..., tag = "p")
}
\arguments{
\item{text}{A character vector of text to add. Each entry will be a separate element. Text can include
links to be embedded, in the form \code{"[text](url)"}.}

\item{...}{Attributes to add to each element.}

\item{tag}{The tag name of each element.}
}
\value{
<template: Description of what is returned.>
}
\description{
Adds regular text elements to a website.
}
\examples{
\dontrun{
# regular text
page_text("text to be added")
}
}
