% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{phi}
\alias{phi}
\title{phi}
\usage{
phi(phi_s_l, phi_r_r, verbose = FALSE)
}
\arguments{
\item{phi_s_l}{numeric \link{0,1}: probability of interaction of a ligand with its counterpart receptor based on the relative active fraction of this ligand in the sending cell type.}

\item{phi_r_r}{numeric \link{0,1}: probability of interaction of a receptor with its counterpart ligand based on the relative active fraction of this receptor in the receiving cell type.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
numeric \link{0,1}: probability of interaction based on the relative mean expression of the ligand and receptor in the sending and receiving cell types.
}
\description{
calculates the active fraction based probability of communication
}
