% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_cellType_gene.R
\name{phi_cellType_gene}
\alias{phi_cellType_gene}
\title{phi_cellType_gene}
\usage{
phi_cellType_gene(a_cellType_gene, a_cellType_gene_max, verbose = FALSE)
}
\arguments{
\item{a_cellType_gene}{numeric \link{0,1}: active fraction data for the ligand (or receptor) in the cell type of interest.}

\item{a_cellType_gene_max}{numeric \link{0,1}: maximum active fraction for the ligand (or receptor) in the cell type of interest across all samples.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
a_norm: numeric \link{0,1}: normalised active fraction for the ligand (or receptor) in the cell type of interest.
}
\description{
calculates the relative active fraction of a ligand or receptor based on its expression level in the cell type of interest.
}
\examples{
a_cellType_gene <- 0.1
a_cellType_gene_max <- 0.1
phi_cellType_gene(a_cellType_gene = a_cellType_gene
                 ,a_cellType_gene_max = a_cellType_gene_max
                ,verbose = TRUE
)
}
