% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_forests.R
\name{plot_all_forests}
\alias{plot_all_forests}
\title{plot_all_forests}
\usage{
plot_all_forests(
  my_idx,
  my_anno_interactions,
  keep_order = FALSE,
  show_labels = FALSE,
  plot_legend = TRUE
)
}
\arguments{
\item{my_idx}{Index of the interactions to be plotted}

\item{my_anno_interactions}{dataframe: Dataframe containing the interactions, having the rows as interactions and columns as the different parameters and the interaction_ID.}

\item{keep_order}{a logical value. If TRUE then keeps oder of my_anno_interactions ignoring the my_idx, default is FALSE}

\item{show_labels}{a logical value. If true interaction_IDs are plotted, default is FALSE}

\item{plot_legend}{a logical value. If true the legend is plotted, default is TRUE}
}
\value{
a list of 7 forest plots
}
\description{
creates 7 forest plots for different parameters.
}
\examples{
# plot_all_forests
plot_all_forests(my_idx
               ,my_anno_interactions
)

}
