% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumW.R
\name{plot_cumW}
\alias{plot_cumW}
\title{plot_cumW}
\usage{
plot_cumW(df, threshold_log10_cum_weight)
}
\arguments{
\item{df:}{dataframe: dataframe containing the interactions, including a column "log10_cum_weight"}

\item{threshold_log10_cum_weight:}{numeric: threshold for log10 cumulative weight of interactions. Interactions with a log10 cumulative weight above this threshold will be considered as passing the threshold.}
}
\value{
histogram plot
}
\description{
creates a histogram plot of log10 cumulative weight of interactions.
}
\examples{
# plot the cumulative weight of the interactions
plot_cumW(df = anno_interactions_allSamples[[1]], threshold_log10_cum_weight = 0)

}
