% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumW_vs_nrSamples.R
\name{plot_cumW_vs_nrSamples}
\alias{plot_cumW_vs_nrSamples}
\title{plot_cumW_vs_nrSamples}
\usage{
plot_cumW_vs_nrSamples(
  df,
  threshold_log10_cum_weight,
  threshold_nr_expr_samples
)
}
\arguments{
\item{df:}{dataframe: dataframe containing the interactions, including columns "log10_cum_weight" and "nr_expr_samples"}

\item{threshold_log10_cum_weight:}{numeric: threshold for log10 cumulative weight of interactions. Interactions with a log10 cumulative weight above this threshold will be considered as passing the threshold.}

\item{threshold_nr_expr_samples:}{numeric: threshold for the number of samples in which the interactions are expressed. Interactions expressed in more than this number of samples will be considered as passing the threshold.}
}
\value{
list of ggplots: list containing the scatter plot, the y-density plot, and the x-density plot
}
\description{
creates a scatter plot of log10 cumulative weight of interactions vs number of samples in which the interactions are expressed. It also gives a visual representation of whether the interactions passed a certain threshold for log10 cumulative weight and number of samples expressed in.
}
\examples{
# plot the cumulative weight vs number of samples in which the interactions are expressed
plot_cumW_vs_nrSamples(df = anno_interactions_allSamples[[1]]
                      ,threshold_log10_cum_weight = 0
                     ,threshold_nr_expr_samples = 0
)
}
