% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest.R
\name{plot_forest}
\alias{plot_forest}
\title{plot_forest}
\usage{
plot_forest(my_df, my_title, plot_legend = TRUE, min, max)
}
\arguments{
\item{my_df:}{a data frame containing the following columns: 'interaction_ID' and 'log2FC'. The interaction_ID column must be unique.}

\item{my_title:}{character: title for the plot}

\item{plot_legend:}{logical: should the legend be plotted (default is TRUE)}

\item{min:}{numeric: minimal log2FC value to be plotted (default is min of my_df$log2FC)}

\item{max:}{numeric: maximal log2FC value to be plotted (default is max of my_df$log2FC)}
}
\value{
ggplot object
}
\description{
creates a horizontal bar chart of log2FC expression values for a set of interactions. The function also allows to customize the color scale and the range of the log2FC values.
}
\examples{
# plot_forest
plot_forest(my_df
          ,my_title
         ,plot_legend = TRUE
        ,min
       ,max
)

}
