% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fracSamples.R
\name{plot_fracSamples}
\alias{plot_fracSamples}
\title{plot_fracSamples}
\usage{
plot_fracSamples(df, threshold_frac_samples_per_condition, ...)
}
\arguments{
\item{df:}{dataframe: dataframe containing the interactions, including columns "frac_samples_controls" and "frac_samples_cases"}

\item{threshold_frac_samples_per_condition:}{numeric: threshold for the fraction of samples in which the interactions are detected.}
}
\value{
scatter plot, x axis is fraction of samples in controls, y axis is fraction of samples in cases
}
\description{
creates a scatter plot of fraction of samples in which the interactions are expressed. It also gives a visual representation of whether the interactions passed a certain threshold for log10 cumulative weight and number of samples expressed in.
}
\examples{
# plot the fraction of samples in which the interactions are detected
plot_fracSamples(df = anno_interactions_allSamples[[1]]
                      ,threshold_frac_samples_per_condition = 0.8
)
}
