% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(
  comm_result,
  which_interactions = NULL,
  by_param,
  values_to_plot,
  row_font_size = 8,
  color_case = "#7C001F",
  color_control = "#7AC5CD",
  color_values = NULL
)
}
\arguments{
\item{comm_result}{A dataframe of common interactions}

\item{which_interactions}{A logical vector with length of all interactions, indicating which interactions to include in the heatmap. NULL if top and by_param are set or "all"}

\item{by_param}{A string indicating which column of the common interactions dataframe to use for filtering interactions, e.g. "passed_QC_filter", "passed_FDR_threshold", "passed_log2FC_threshold", "sign"}

\item{values_to_plot}{A string indicating which column of the common interactions dataframe to use as the values in the heatmap. e.g. "interactions_weights", "expr_l_s_active", "expr_r_r_active", "nr_l_s_active", "nr_r_r_active", "phi", "phi_l_s", "phi_r_r", "p", "p_l_s", "p_r_r"}

\item{row_font_size}{Font size for the row labels}

\item{color_case}{A string indicating the color for case samples in the heatmap}

\item{color_control}{A string indicating the color for control samples in the heatmap}
}
\description{
plots heatmap of chosen parameter
}
\examples{
# load example data
data("comm_result")
# calculate general statistics
comm_result <- general_stat(comm_result)
# plot heatmap
plot_heatmap(comm_result = comm_result
            ,which_interactions = NULL
           ,by_param = "passed_QC_filter"
          ,values_to_plot = "interactions_weights"
        ,row_font_size = 8
     ,color_case = "#7C001F" # "darkred"
 ,color_control = "#7AC5CD" # "CadetBlue3"
)

}
