% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_log2FCw_graph.R
\name{plot_mean_log2FCw_graph}
\alias{plot_mean_log2FCw_graph}
\title{plot_mean_log2FCw_graph}
\usage{
plot_mean_log2FCw_graph(my_interactions, my_idx, direction)
}
\arguments{
\item{my_interactions}{A list containing information about the interactions, samples, and their annotations.}

\item{my_idx}{A logical vector of the same length as number of interactions in my_interactions. Only interactions with TRUE values will be plotted.}

\item{direction}{A character string, "down" or "up", indicating the direction of the plot (up=upregulate, down=downregulated).}
}
\value{
A graph of cell types interactions.
}
\description{
creates a network graph of cell types interactions, where each interactions represents the mean of log2FC weights of interactions between two cell types.
}
\examples{
# plot_mean_log2FCw_graph
plot_mean_log2FCw_graph(my_interactions
                 ,my_idx
                ,"up"
)

}
