% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDB.R
\name{process_adhesive_DB}
\alias{process_adhesive_DB}
\title{Process Adhesive Ligand-Receptor Database}
\usage{
process_adhesive_DB(complete, annotation, ligand_list = list(), receptor_list = list())
}
\arguments{
\item{complete}{A data frame representing the complete set of ligand-receptor interactions.}

\item{annotation}{A data frame containing annotations for ligands and receptors.}

\item{ligand_list}{A list of gene symbols representing ligands. Optional, used for additional annotations.}

\item{receptor_list}{A list of gene symbols representing receptors. Optional, used for additional annotations.}
}
\value{
A processed data frame of the adhesive ligand-receptor database, annotated and filtered based on various criteria including gene family and interaction directionality.
}
\description{
This function processes an adhesive database. It involves filtering, rearranging, and annotating the database based on gene families and interaction properties. This includes handling of directional and non-directional interactions, as well as manual annotations for specific gene families.
}
\examples{
# Assuming 'complete' and 'annotation' are previously prepared data frames
adhesive_db <- process_adhesive_DB(complete, annotation)

# With additional ligand and receptor lists
adhesive_db <- process_adhesive_DB(complete, annotation, ligand_list, receptor_list)

}
