% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDB.R
\name{process_binary_pairs}
\alias{process_binary_pairs}
\title{Process Binary Pairs}
\usage{
process_binary_pairs(both_db, pt_interactions)
}
\arguments{
\item{both_db}{A data frame of combined ligand-receptor interactions.}

\item{pt_interactions}{A data frame of post-translational interactions.}
}
\value{
A data frame of merged binary pairs and interactions filtered through PPI, without duplicates.
}
\description{
This function processes the single component interactions from a combined ligand-receptor database and merges them with PPI interactions.
}
\examples{
both_db <- import_db("both")
pairwise_pairs <- create_pairwise_pairs(both_db)
pt_interactions <- filter_pairs_with_ppi(pairwise_pairs)
# Assuming 'both_db' and 'pt_interactions' are available data frames
complete_interactions <- process_binary_pairs(both_db, pt_interactions)

}
