% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDB.R
\name{process_lr_db}
\alias{process_lr_db}
\title{Process Ligand-Receptor Database}
\usage{
process_lr_db(complete, annotation, all_ligands, all_receptors)
}
\arguments{
\item{complete}{Database.}

\item{annotation}{A data frame with gene annotations.}

\item{all_ligands}{(optional) A vector of all ligand genesymbols.}

\item{all_receptors}{(optional) A vector of all receptor genesymbols.}
}
\value{
A data frame of the processed ligand-receptor pairs with correct ligand-receptor directions.
}
\description{
This function processes a ligand-receptor database, ensuring correct ligand-receptor direction based on provided annotation.
}
\examples{
# Assuming 'complete', 'annotation', 'all_ligands', and 'all_receptors' are available
processed_db <- process_lr_db(complete, annotation, all_ligands, all_receptors)

}
