% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{rescale}
\usage{
rescale(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector with all values rescaled to be between 0 and 1
}
\description{
Rescales a numeric vector to range between 0 and 1
}
