% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho}
\alias{rho}
\title{rho}
\usage{
rho(rho_s, rho_r, verbose = FALSE)
}
\arguments{
\item{rho_s}{numeric \link{0,1}: normalized abundance of the sending cell type within the sample of interest.}

\item{rho_r}{numeric \link{0,1}: normalized abundance of the receiving cell type within the sample of interest.}

\item{verbose}{Logical, whether to print additional information}
}
\value{
rho numeric \link{0,1}: the normalized abundance component of the weight formula
}
\description{
calculates the normalized abundance component of the weight formula.
}
