% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_cellType.R
\name{rho_cellType}
\alias{rho_cellType}
\title{rho_cellType}
\usage{
rho_cellType(f_cellType, f_cellType_max, verbose = FALSE)
}
\arguments{
\item{f_cellType}{numeric \link{0,1}: abundance of the cell types (as fractions).}

\item{f_cellType_max}{numeric \link{0,1}: maximum abundance of the cell type of interest across all samples.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
relative abundance for each cell type.
}
\description{
calculates relative abundance of cell types.
}
