% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_build.R
\name{site_build}
\alias{site_build}
\title{Render a Website}
\usage{
site_build(dir, file = "site.R", outdir = "docs", name = "index.html",
  variables = NULL, options = list(), bundle_data = FALSE,
  open_after = FALSE, aggregate = TRUE, force = FALSE)
}
\arguments{
\item{dir}{Path to the site project directory.}

\item{file}{Name of the R file to build the site from.}

\item{outdir}{Subdirectory of \code{dir} in which to place the built site files (where the site is to be
served from).}

\item{name}{Name of the HTML file to be created.}

\item{variables}{A character vector of variable names to include from the data. If no specified,
all variables are included.}

\item{options}{A list with options to be passed to the site. These will be written to \code{docs/settings.json},
which can be edited by hand.}

\item{bundle_data}{Logical; if \code{TRUE}, will write the data to the site file; useful when
running the site locally without server. Otherwise, the data will be loaded separately through an http request.}

\item{open_after}{Logical; if \code{TRUE}, will open the site in a browser after it is built.}

\item{aggregate}{Logical; if \code{TRUE}, and there is a larger datasets with IDs that partially match
IDs in a smaller dataset or that has a map to those IDs, and there are NAs in the smaller dataset, will
attempt to fill NAs with averages from the larger dataset.}

\item{force}{Logical; if \code{TRUE}, will reprocess data even if the source data is older than the existing
processed version.}
}
\value{
Invisible path to the written file.
}
\description{
Write HTML output from the \code{site.R} file in a site project.
}
\examples{
\dontrun{
# run from within a site project directory, initialized with `init_site()`
site_build(".")

# bundle data to most easily run locally
site_build(".", bundle_data = TRUE)
}
}
\seealso{
To initialize a site project, use \code{\link{init_site}}.
}
