% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_diff.R
\name{test_diff}
\alias{test_diff}
\title{test_diff}
\usage{
test_diff(
  comm_result,
  which_adjustment = "fdr",
  threshold_fdr = 0.1,
  threshold_log2FC = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{which_adjustment}{method used for multiple testing correction by the p.adjust function. The default is 'fdr'.}

\item{threshold_fdr}{the p-value threshold for adjusted p-values. The default value is 0.1}

\item{threshold_log2FC}{the Log2 fold-change threshold for significant interactions. The default value is 1.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}

\item{...}{Additional parameters passed to the t.test or the wilcoxon test function}

\item{comm_result:}{list containing information about the interactions, samples, and their annotations.}
}
\value{
A comm_result object with the following columns added:
\itemize{
\item anno_interactions:
-- log2FC_weights: numeric: log2 fold change of the mean interaction weight between the cases and the controls
-- p.value: numeric: p -value
-- p.adj: numeric: adjusted p -value
-- passed_log2FC_threshold: a logical value indicating whether an interaction passed the passed_log2FC_threshold
-- passed_FDR_threshold: a logical value indicating whether an interaction passed the passed_FDR_threshold
-- sign: a logical value indicating whether an interaction passed both the passed_log2FC_threshold and the passed_FDR_threshold thresholds
\item thresholds:
-- threshold_log2FC: numeric: the Log2 fold-change threshold for significant interactions.
-- threshold_fdr: numeric: the p-value threshold for adjusted p-values.
}
}
\description{
tests if an interaction is statistically differential between the cases and the controls.
}
\examples{
# load example data
data("comm_result")
# calculate general statistics
comm_result <- general_stat(comm_result)
# test for differential expression
comm_result <- test_diff(comm_result
                       ,which_adjustment = "fdr"
                      ,verbose = TRUE
                     ,threshold_fdr = 0.1
                   ,threshold_log2FC = 1
)

}
