% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.R
\name{w}
\alias{w}
\title{w}
\usage{
w(rho, phi, p, verbose = FALSE)
}
\arguments{
\item{rho}{numeric \link{0,1}: relative fraction of cell types.}

\item{phi}{numeric \link{0,1}: probability of interaction between two cell types based on fraction of communication cells.}

\item{p}{numeric \link{0,1}: probability of interaction between two cell types based on expression levels of the ligand and receptor.}

\item{verbose}{a logical value indicating whether to print progress messages (default: FALSE).}
}
\value{
numeric \link{0,1}: weight of the interaction.
}
\description{
calculates the weight of the interactions between two nodes based on fraction of cells and expression based probability of interaction.
}
