% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community-package.R
\docType{package}
\name{community-package}
\alias{community}
\alias{community-package}
\title{community: Community Analysis Framework}
\description{
A toolset and web framework to help manage and present data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/uva-bi-sdad/community}
  \item Report bugs at \url{https://github.com/uva-bi-sdad/community/issues}
}

}
\author{
\strong{Maintainer}: Micah Iserman \email{micah.iserman@gmail.com}

Other contributors:
\itemize{
  \item Biocomplexity Institute [copyright holder, funder]
  \item Alan Wang [contributor]
  \item Prashanth Wagle [contributor]
}

}
\keyword{internal}
