% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community-package.R
\docType{package}
\name{community-package}
\alias{community}
\alias{community-package}
\title{community: Community Analysis Framework}
\description{
Offers a framework and toolset for the analysis of communities. Helps (1a) assess, clean, and unify diverse data sources relating to a community, or (1b) simulate a community, then (2) calculate a range of measures of that community, and (3) present and disseminate final data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/miserman/community}
  \item Report bugs at \url{https://github.com/miserman/community/issues}
}

}
\author{
\strong{Maintainer}: Micah Iserman \email{micah.iserman@gmail.com}

Authors:
\itemize{
  \item Biocomplexity Institute [copyright holder]
}

}
\keyword{internal}
