% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_add.R
\name{data_add}
\alias{data_add}
\title{Adds documentation of a dataset to a datapackage}
\usage{
data_add(filename, meta = list(), packagename = "datapackage.json",
  dir = ".", write = TRUE, refresh = TRUE, sha = "512",
  clean = FALSE, open_after = FALSE)
}
\arguments{
\item{filename}{A character vector of paths to plain-text tabular data files, relative to \code{dir}.}

\item{meta}{Information about each data file. A list with a list entry for each entry in
\code{filename}; see details. If a single list is provided for multiple data files, it will apply to all.}

\item{packagename}{Package to add the metadata to; path to the \code{.json} file relative to
\code{dir}, or a list with the read-in version.}

\item{dir}{Directory in which to look for \code{filename}, and write \code{packagename}.}

\item{write}{Logical; if \code{FALSE}, returns the \code{paths} metadata without reading or rewriting
\code{packagename}.}

\item{refresh}{Logical; if \code{FALSE}, will retain any existing dataset information.}

\item{sha}{A number specifying the Secure Hash Algorithm function,
if \code{openssl} is available (checked with \code{Sys.which('openssl')}).}

\item{clean}{Logical; if \code{TRUE}, strips special characters before saving.}

\item{open_after}{Logical; if \code{TRUE}, opens the written datapackage after saving.}
}
\value{
An invisible version of the updated datapackage, which is also written to
\code{datapackage.json} if \code{write = TRUE}.
}
\description{
Add information about variables in a dataset to a \code{datapackage.json} metadata file.
}
\details{
\code{meta} should be a list with unnamed entries for entry in \code{filename},
and each entry can include a named entry for any of these:
\describe{
\item{source}{
A list or list of lists with entries for at least \code{name}, and ideally for \code{url}.
}
\item{ids}{
A list or list of lists with entries for at least \code{variable} (the name of a variable in the dataset).
Might also include \code{map} with a list or path to a JSON file resulting in a list with an
entry for each ID, and additional information about that entity, to be read in a its features.
All files will be loaded to help with aggregation, but local files will be included in the datapackage,
whereas hosted files will be loaded client-side.
}
\item{time}{
A string giving the name of a variable in the dataset representing a repeated observation of the same entity.
}
\item{variables}{
A list with named entries providing more information about the variables in the dataset.
See \code{\link{data_measure_info}}.
}
}
}
\examples{
\dontrun{
# write example data
write.csv(mtcars, "mtcars.csv")

# add it to an existing datapackage.json file in the current working directory
data_add("mtcars.csv")
}
}
\seealso{
Initialize the \code{datapackage.json} file with \code{\link{init_data}}.
}
