% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{init_data}
\alias{init_data}
\title{Create a datapackage.json template}
\usage{
init_data(name, title = name, dir = ".", ..., write = TRUE,
  overwrite = FALSE, quiet = !interactive())
}
\arguments{
\item{name}{A unique name for the dataset; allowed characters are \code{[a-z._/-]}.}

\item{title}{A display name for the dataset; if not specified, will be a formatted version of \code{name}.}

\item{dir}{Directory in which to save the \code{datapackage.json} file.}

\item{...}{passes arguments to \code{\link{data_add}}.}

\item{write}{Logical; if \code{FALSE}, the package object will not be written to a file.}

\item{overwrite}{Logical; if \code{TRUE} and \code{write} is \code{TRUE}, an existing
\code{datapackage.json} file will be overwritten.}

\item{quiet}{Logical; if \code{TRUE}, will not print messages or navigate to files.}
}
\value{
An invisible list with the content written to the \code{datapackage.json} file.
}
\description{
Initialize dataset documentation with a \code{datapackage.json} template, based on a
\href{https://specs.frictionlessdata.io/data-package}{Data Package} standard.
}
\examples{
\dontrun{
# make a template datapackage.json file in the current working directory
init_data("mtcars", "Motor Trend Car Road Tests")
}
}
\seealso{
Add basic information about a dataset with \code{\link{data_add}}.
}
