% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_datacommons.R
\name{init_datacommons}
\alias{init_datacommons}
\title{Initialize a Data Commons}
\usage{
init_datacommons(dir, name = "data commons", repos = NULL,
  default_user = "", overwrite = FALSE, verbose = interactive())
}
\arguments{
\item{dir}{Path to the desired data commons directory.}

\item{name}{Name of the data commons.}

\item{repos}{A vector of repository names to add to \code{commons.json}.}

\item{default_user}{GitHub username to prepend to repository names if needed.}

\item{overwrite}{Logical; if \code{TRUE}, will overwrite existing datacommons files in \code{dir}.}

\item{verbose}{Logical; if \code{FALSE}, suppresses messages.}
}
\value{
Path to the datacommons directory.
}
\description{
Initialize a project to keep track of separate dataset repositories and distributions.
}
\examples{
\dontrun{
init_datacommons("../datacommons")
}
}
