% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_site.R
\name{init_site}
\alias{init_site}
\title{Create a website}
\usage{
init_site(dir, title = "app", with_data = TRUE, overwrite = FALSE,
  quiet = !interactive())
}
\arguments{
\item{dir}{Directory in which to create the site's structure. Will be created if it does not exist.}

\item{title}{Title of the site.}

\item{with_data}{logical; if \code{FALSE}, a data sub-directory and package will not be created.}

\item{overwrite}{logical; if \code{TRUE}, will overwrite existing site files in \code{dir}.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages and does not navigate to the file when finished.}
}
\value{
Path to the created site directory.
}
\description{
Create a repository for a static website for data documentation and exploration.
}
\examples{
\dontrun{
# initialize site in the current working directory
init_site(".")
}
}
