% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_button.R
\name{input_button}
\alias{input_button}
\title{Add a button to a website}
\usage{
input_button(label, target = "reset_selection", id = label, class = "",
  ..., note = NULL)
}
\arguments{
\item{label}{Text in the button for the user.}

\item{target}{The ID of an input element to reset, or the name of a function to trigger. Available function are
\code{refresh} (to force a reprocessing of the current input state), \code{reset_selection} (to reset all input
elements that are not settings), and \code{reset_storage} (to clear local storage, reset all settings, and reload).
This can also be a URL (including protocol) for a button-like link.}

\item{id}{Unique ID of the element to be created.}

\item{class}{Additional class names to add to the element.}

\item{...}{Additional attributes to set on the element.}

\item{note}{Text to display as a tooltip for the button.}
}
\value{
A character vector of the contents to be added.
}
\description{
Adds a button to reset an input or trigger a predefined function to a website.
}
\examples{
\dontrun{
input_button("Reset")
}
}
\seealso{
For buttons to select between a set of options, use \code{\link{input_buttongroup}}.
}
