% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_select.R
\name{input_select}
\alias{input_select}
\title{Add a select input to a website}
\usage{
input_select(label, options, default = -1, display = options, id = label,
  variable = NULL, dataset = NULL, depends = NULL,
  reset_button = FALSE)
}
\arguments{
\item{label}{Label of the input for the user.}

\item{options}{A vector of options, the name of a variable from which to pull levels, or either \code{"datasets"}
or \code{"variables"} to select names of datasets or variables.}

\item{default}{Which of the options to default to; either its index or value.}

\item{display}{A display version of the options.}

\item{id}{Unique id of the element to be created.}

\item{variable}{The name of a variable from which to get levels (overwritten by \code{depends}).}

\item{dataset}{The name of an included dataset, where \code{variable} should be looked for; only applies when
there are multiple datasets with the same variable name.}

\item{depends}{The id of another input on which the options depend; this will take president over \code{dataset}
and \code{variable}, depending on this type of input \code{depends} points to.}

\item{reset_button}{If specified, adds a button after the select element that will revert the selection
to its default; either \code{TRUE}, or text for the reset button's label.}
}
\value{
A character vector of the contents to be added.
}
\description{
Adds an input to select from the entered options.
}
\examples{
\dontrun{
input_select()
}
}
