% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_text.R
\name{input_text}
\alias{input_text}
\title{Add a text input to a website}
\usage{
input_text(id, label = id, ..., default = NULL, note = NULL,
  multiline = FALSE, class = NULL, floating_label = TRUE)
}
\arguments{
\item{id}{Unique ID of the element to be created.}

\item{label}{Label of the input for the user.}

\item{...}{Other attributes to add to the input.}

\item{default}{Default value of the input, which will appear as a placeholder.}

\item{note}{Text to display as a tooltip for the input or textarea.}

\item{multiline}{Logical; if \code{TRUE}, create a \code{textarea} element, instead of an \code{input} element
to accept multiple lines of text.}

\item{class}{Class names to add to the input's list.}

\item{floating_label}{Logical; if \code{FALSE}, labels are separate from their input elements.}
}
\value{
A character vector of the contents to be added.
}
\description{
Adds an direct text input element to a website.
}
\examples{
\dontrun{
input_text("entered_text", "Enter Text:")
}
}
