% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_R_Database.R
\docType{data}
\name{LR_database}
\alias{LR_database}
\title{Ligand Receptor Database}
\format{
Ligand Receptor Database, each row is a ligand-receptor pair.
}
\usage{
data(LR_database)
}
\description{
Data from previous studies of the ligand-receptor interaction.
}
\examples{
data(LR_database)
head(LR_database)
}
\references{
iTALK: An Integrated Tool for Ligand-Receptor Interaction Analysis DB
}
\keyword{datasets}
