% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_credits.R
\name{output_credits}
\alias{output_credits}
\title{Adds a legend to a website}
\usage{
output_credits(add = NULL, exclude = NULL)
}
\arguments{
\item{add}{A list of credits to add. Each credit should be a list with at least entries for \code{"name"}
and \code{"url"}, and optionally a \code{"version"} and/or \code{"description"}.}

\item{exclude}{Names of automatic credits to exclude. The automatic credits are \code{"bootstrap"},
\code{"leaflet"} (from \code{\link{output_map}}), \code{"plotly"} (from \code{\link{output_plot}}), and
\code{"datatables"} (from \code{\link{output_table}}).}
}
\value{
A character vector of the contents to be added.
}
\description{
Adds a legend based on a specified color palette.
}
\examples{
\dontrun{
output_legend("reds")
}
}
