% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_info.R
\name{output_info}
\alias{output_info}
\title{Add a tooltip-like display to a website}
\usage{
output_info(title = "Overall", message = "", dataview = NULL, subto = NULL)
}
\arguments{
\item{title}{Default title at the top of the element.}

\item{message}{Default message below the title, to be replaced by the information table.}

\item{dataview}{The ID of a dataview, used for the persistent display.}

\item{subto}{A vector of output IDs to receive hover events from.}
}
\value{
A character vector of the content to be added.
}
\description{
Adds an output to display information about a hovered-over or selected data point.
}
\examples{
\dontrun{
output_info("Initial View", "Hover over plot elements for more information.")
}
}
