% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_info.R
\name{output_info}
\alias{output_info}
\title{Add a tooltip-like display to a website}
\usage{
output_info(title = "features.name", body = "value",
  variable_info = list(title = "short_name", body = "short_description"),
  default = c(title = "Overall", body = "No selection."), dataview = NULL,
  subto = NULL)
}
\arguments{
\item{title}{Title text, or the source of title text.}

\item{body}{A list of entries in the info body section, which can be raw text or references to features or
data variables. Named entries appear in two columns, with the name in the first.}

\item{variable_info}{A list with entries for \code{"title"} and \code{"body"}, which can be raw text or reference
to variable metadata entries.}

\item{default}{A list with entries for \code{"title"} and \code{"body"}, which are treated as raw text.}

\item{dataview}{The ID of a dataview, used for the persistent display.}

\item{subto}{A vector of output IDs to receive hover events from.}
}
\value{
A character vector of the content to be added.
}
\description{
Adds an output to display information about a hovered-over or selected data point.
}
\examples{
\dontrun{
output_info("Initial View", "Hover over plot elements for more information.")
}
}
