% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_legend.R
\name{output_legend}
\alias{output_legend}
\title{Adds a legend to a website}
\usage{
output_legend(palette = "divergent", lower = "Lower", center = "",
  higher = "Higher", class = "")
}
\arguments{
\item{palette}{Name of an included color palette (from
\href{https://colorbrewer2.org}{colorbrewer}); one of \code{"divergent"} (default), \code{"reds"},
\code{"greens"}, \code{"grays"}, or \code{"cat"}.}

\item{lower}{Text to appear at the beginning of the legend, labeling smaller.}

\item{center}{Text to appear at the midpoint of the legend, labeling central values.}

\item{higher}{Text to appear at the end of the legend, labeling larger values.}

\item{class}{Class names to add to the legend table.}
}
\value{
A character vector of the contents to be added.
}
\description{
Adds a legend based on a specified color palette.
}
\examples{
\dontrun{
output_legend("reds")
}
}
