% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_legend.R
\name{output_legend}
\alias{output_legend}
\title{Adds a legend to a website}
\usage{
output_legend(palette = "", variable = NULL, dataview = NULL,
  id = NULL, click = NULL, subto = NULL, class = "", show_na = TRUE)
}
\arguments{
\item{palette}{Name of an included color palette, or palette selection input;
for discrete scales, one of \code{"rdylbu7"}, \code{"orrd7"}, \code{"gnbu7"},
\code{"brbg7"}, \code{"puor7"}, \code{"prgn6"}, \code{"reds5"}, \code{"greens5"}, \code{"greys4"}, \code{"paired4"} (from
\href{https://colorbrewer2.org}{colorbrewer}); for continuous scales, one of \code{"grey"}, \code{"brown"}, \code{"purple"},
\code{"prgn"}, \code{"puor"}, \code{"rbbu"}, \code{"prgn"}, \code{"vik"} (default), or \code{"lajolla"}.}

\item{variable}{Name of a variable or ID of a variable selector to display values of. Defaults to
the \code{y} variable of \code{dataview} if one is specified.}

\item{dataview}{The ID of an \code{\link{input_dataview}} component.}

\item{id}{Unique ID of the legend element.}

\item{click}{The ID of an input to set to an entity's ID near the current cursor location on the current scale.}

\item{subto}{A vector of output IDs to receive hover events from.}

\item{class}{Class names to add to the legend element.}

\item{show_na}{Logical; if \code{FALSE}, does not add the separate section showing the color of missing values.}
}
\value{
A character vector of the contents to be added.
}
\description{
Adds a legend based on a specified color palette.
}
\examples{
output_legend()
}
