% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_legend.R
\name{output_legend}
\alias{output_legend}
\title{Adds a legend to a website}
\usage{
output_legend(palette = "", lower = "Lower", higher = "Higher",
  dataview = NULL, id = NULL, subto = NULL, class = "")
}
\arguments{
\item{palette}{Name of an included color palette (from
\href{https://colorbrewer2.org}{colorbrewer}); one of \code{"rdylbu7"} (default), \code{"orrd7"}, \code{"gnbu7"},
\code{"brbg7"}, \code{"puor7"}, \code{"prgn6"}, \code{"reds5"}, \code{"greens5"}, \code{"greys4"}, \code{"paired4"}.}

\item{lower}{Text to appear at the beginning of the legend, labeling smaller.}

\item{higher}{Text to appear at the end of the legend, labeling larger values.}

\item{dataview}{The ID of an \code{\link{input_dataview}} component.}

\item{id}{Unique ID of the legend element.}

\item{subto}{A vector of output IDs to receive hover events from.}

\item{class}{Class names to add to the legend element.}

\item{center}{Text to appear at the midpoint of the legend, labeling central values.}
}
\value{
A character vector of the contents to be added.
}
\description{
Adds a legend based on a specified color palette.
}
\examples{
output_legend()
}
