% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_table.R
\name{output_table}
\alias{output_table}
\title{Add a table to a webpage}
\usage{
output_table(variables = NULL, dataset = NULL, dataview = NULL,
  options = NULL, filters = NULL, wide = FALSE, single_variable = TRUE,
  class = "compact")
}
\arguments{
\item{variables}{Variable names in the included data, or the ID of a
variable selector to show in columns of the table.}

\item{dataset}{The name of a dataset, or ID of a dataset selector, to find \code{variables} in;
used if \code{dataview} is not specified.}

\item{dataview}{The ID of an \code{\link{input_dataview}} component.}

\item{options}{A list of configuration options, see
\href{https://datatables.net/reference/option}{DataTables Documentation}.}

\item{filters}{A list with names of \code{meta} entries (from \code{variable} entry in \code{\link{data_add}}'s
\code{meta} list), and values of target values for those entries, or the IDs of value selectors.}

\item{wide}{Logical; if \code{TRUE}, multiple variables are shown in separate columns.}

\item{single_variable}{Logical; if \code{TRUE} (default), \code{variables} is treated as a single variable,
or source of a single variable, and time is displayed as columns. Otherwise, if a selector is specified, its
options are treated as the set of variables.}

\item{class}{Class names to add to the table.}
}
\value{
A character vector of the content to be added.
}
\description{
Adds a DataTables table to a webpage, based on specified or selected .
}
\examples{
\dontrun{
output_table()
}
}
