% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_navbar.R
\name{page_navbar}
\alias{page_navbar}
\title{Add a navbar (header) section to a website}
\usage{
page_navbar(title = "", logo = "", ...)
}
\arguments{
\item{title}{A title to appear in the header.}

\item{logo}{URL of an image to appear in the header.}

\item{...}{Lists specifying submenus and their items. Each list should have entries for
\code{"name"} (the navbar button text), \code{"title"} (the menu title), \code{"placement"}
(start, top, end, or bottom), \code{"backdrop"} (logical), and \code{items} (a list of menu items).}
}
\value{
A character vector of the content to be added.
}
\description{
Adds a bar at the top of the page, potentially with a title and logo image.
}
\examples{
page_navbar(
  "Site Name",
  submenus = list(
    name = "menu",
    title = "Menu",
    items = list(input_select("menu select", c("a", "b")))
  )
)
}
