% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_navbar.R
\name{page_navbar}
\alias{page_navbar}
\title{Add a navbar (header) section to a website}
\usage{
page_navbar(..., title = "", logo = "", breakpoint = "md",
  logo.height = "24px", logo.width = "24px")
}
\arguments{
\item{...}{Content to add to the navbar. Can be lists specifying submenus and their items.
Each list can have entries for \code{"name"} (the navbar button text), \code{"title"} (the menu title),
\code{"foot"} (the menu footer), \code{"id"} (the ID of the menu element),
\code{"class"} (class of the menu element), \code{"placement"} (start, top, end, or bottom),
\code{"backdrop"} (logical, where \code{FALSE} disables the backdrop and makes the menu shift
main content), and \code{items} (a list of menu items).}

\item{title}{A title to appear in the header.}

\item{logo}{URL of an image to appear in the header.}

\item{breakpoint}{Bootstrap breakpoint code specifying the width at which the menu collapses;
default is \code{"md"}.}

\item{logo.height}{Character or number setting the height of the logo.}

\item{logo.width}{Character or number setting the width of the logo.}
}
\value{
A character vector of the content to be added.
}
\description{
Adds a bar at the top of the page, potentially with a title and logo image.
}
\examples{
page_navbar(
  title = "Site Name",
  list(
    name = "menu",
    title = "Menu",
    items = list(input_select("menu select", c("a", "b")))
  )
)
}
