% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_build.R
\name{site_build}
\alias{site_build}
\title{Render a Website}
\usage{
site_build(dir = ".", file = "site.r", outdir = "docs",
  name = "index.html", bundle_data = FALSE, open_after = FALSE,
  force = FALSE)
}
\arguments{
\item{dir}{Path to the site project directory.}

\item{file}{Name of the R file to build the site from.}

\item{outdir}{Path to place the .}

\item{name}{Path to the site project directory.}

\item{bundle_data}{Logical; if \code{TRUE}, will write the data to the site file; useful when
running the site locally without server. Otherwise, the data will be loaded separately through an http request.}

\item{open_after}{Logical; if \code{TRUE}, will open the site in a browser after it is built.}

\item{force}{Logical; if \code{TRUE}, will reprocess data even if the source data is older than the existing
processed version.}
}
\value{
Invisible path to the written file.
}
\description{
Write HTML output from the \code{site.R} file in a site project.
}
\examples{
\dontrun{
# run from within a site project directory, initialized with `init_site()`
site_build()
}
}
\seealso{
To initialize a site project, use \code{\link{init_site}}.
}
