% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_make_child.R
\name{site_make_child}
\alias{site_make_child}
\title{Make a Copy of a Data Site}
\usage{
site_make_child(parent, dir, update = TRUE, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{parent}{Directory or GitHub repository name of the existing site to be copied.}

\item{dir}{Directory of the child site to put copies in.}

\item{update}{Logical; if \code{TRUE}, replaces existing site files if they are older than existing
files (from a local directory). Defaults to \code{FALSE} if \code{parent} is not a local directory.}

\item{overwrite}{Logical; if \code{TRUE}, overwrites any existing site files.}

\item{quiet}{Logical; if \code{TRUE}, does not send messages.}
}
\value{
Invisible path to the child directory.
}
\description{
Copies baseline files from an existing data site. Useful for making different sites
based on the same data.
}
\examples{
\dontrun{
site_make_child("uva-bi-sdad/community_example", "../community_example")
}
}
